### 2024-11-19 #7501 @editorialbot
[PRE REVIEW]: MRI-NUFFT: Doing non-Cartesian MRI has never been easier
**Submitting author:** @paquiteau (<a href="http://orcid.org/0000-0001-6998-232X">Pierre-Antoine  Comby</a>)
**Repository:** https://github.com/mind-inria/mri-nufft
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
