### 2024-11-19 #7501 @Kevin-Mattheus-Moerman
@paquiteau Just to let you know, we are currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
