### 2024-11-19 175d752
feat: tidy bib
