+++ docs/paper-joss/paper.bib
@@ -1,154 +1,125 @@
 @unpublished{shih_cufinufft_2021,
-  title = {{{cuFINUFFT}}: A Load-Balanced {{GPU}} Library for General-Purpose Nonuniform {{FFTs}}},
-  shorttitle = {{{cuFINUFFT}}},
-  author = {Shih, Yu-hsuan and Wright, Garrett and Andén, Joakim and Blaschke, Johannes and Barnett, Alex H.},
-  date = {2021-03-25},
-  eprint = {2102.08463},
-  eprinttype = {arXiv},
-  eprintclass = {cs, eess, math},
-  url = {http://arxiv.org/abs/2102.08463},
-  urldate = {2021-12-10},
-  abstract = {Nonuniform fast Fourier transforms dominate the computational cost in many applications including image reconstruction and signal processing. We thus present a generalpurpose GPU-based CUDA library for type 1 (nonuniform to uniform) and type 2 (uniform to nonuniform) transforms in dimensions 2 and 3, in single or double precision. It achieves high performance for a given user-requested accuracy, regardless of the distribution of nonuniform points, via cache-aware point reordering, and load-balanced blocked spreading in shared memory. At low accuracies, this gives on-GPU throughputs around 109 nonuniform points per second, and (even including hostdevice transfer) is typically 4–10× faster than the latest parallel CPU code FINUFFT (at 28 threads). It is competitive with two established GPU codes, being up to 90× faster at high accuracy and/or type 1 clustered point distributions. Finally we demonstrate a 5–12× speedup versus CPU in an X-ray diffraction 3D iterative reconstruction task at 10−12 accuracy, observing excellent multi-GPU weak scaling up to one rank per GPU.},
-  langid = {english},
-  keywords = {{Computer Science - Distributed, Parallel, and Cluster Computing},Computer Science - Mathematical Software,Electrical Engineering and Systems Science - Signal Processing,Mathematics - Numerical Analysis,No DOI found},
-  file = {/volatile/home/pc266769/Zotero/storage/K5LLWXZE/shih_cufinufft_2021.pdf}
+  title        = {{{cuFINUFFT}}: A Load-Balanced {{GPU}} Library for General-Purpose Nonuniform {{FFTs}}},
+  shorttitle   = {{cuFINUFFT}},
+  author       = {Shih, Yu-hsuan and Wright, Garrett and Andén, Joakim and Blaschke, Johannes and Barnett, Alex H.},
+  url          = {http://arxiv.org/abs/2102.08463},
+  urldate      = {2021-12-10},
+  date         = {2021-03-25},
+  eprint       = {2102.08463},
+  doi          = {10.48550/arXiv.2102.08463},
+  eprinttype   = {arXiv},
+  eprintclass  = {cs, eess, math},
+  langid       = {english}
 }
-
 @inproceedings{uecker_berkley_2015,
-  title = {Berkley Advanced Reconstruction Toolbox},
-  shorttitle = {Mrirecon/Bart},
-  booktitle = {Proc. {{Intl}}. {{Soc}}. {{Mag}}. {{Reson}}. {{Med}}. 23},
-  author = {Uecker, Martin and Ong, Frank and Tamir, J},
-  date = {2015},
-  location = {Toronto},
-  url = {https://zenodo.org/records/10277939},
-  urldate = {2023-12-19},
-  keywords = {No DOI found},
-  file = {/volatile/home/pc266769/Zotero/storage/LIMD2P5S/10277939.html}
+  title        = {Berkley Advanced Reconstruction Toolbox},
+  shorttitle   = {Mrirecon/Bart},
[…]
