+++ docs/paper-joss/paper.bib
@@ -86,12 +86,15 @@
   booktitle    = {ISMRM Workshop on Data Sampling \& Image Reconstruction},
   note         = {Source code available at https://github.com/mmuckley/torchkbnufft}
 }
-@inproceedings{comby_snake-fmri_2024,
-  title        = {{{SNAKE-fMRI}}: {{A}} Modular {{fMRI}} Simulator from the Space-Time Domain to k-Space Data and Back},
-  author       = {Comby, P.-A. and Vignaud, A. and Ciuciu, P.},
-  booktitle    = {{{ISMRM}} Annual Meeting, (in Press)},
-  location     = {Singapore},
-  date         = 2024
+@misc{comby_snake-fmri_2024,
+  title={SNAKE-fMRI: A modular fMRI data simulator from the space-time domain to k-space and back},
+  author={Pierre-Antoine Comby and Alexandre Vignaud and Philippe Ciuciu},
+  year={2024},
+  eprint={2404.08282},
+  archivePrefix={arXiv},
+  primaryClass={eess.SP},
+  url={https://arxiv.org/abs/2404.08282},
+  doi={10.48550/arXiv.2404.08282}
 }
 @article{farrens_pysap_2020,
   title        = {{{PySAP}}: {{Python Sparse Data Analysis Package}} for Multidisciplinary Image Processing},
@@ -121,5 +124,6 @@
   title        = {{{PySAP-MRI}}: A Python Package for {{MR}} Image Reconstruction},
   author       = {Gueddari, Loubna and Gr, Chaithya and Ramzi, Zaccharie and Farrens, Samuel and Starck, Sophie and Grigis, Antoine and Starck, Jean-Luc and Ciuciu, Philippe},
   year         = 2020,
+  url = {https://inria.hal.science/hal-02399267},
   booktitle    = {{{ISMRM}} Workshop on Data Sampling and Image Reconstruction}
 }
