### 2024-11-19 6e04a36
feat: add CONTRIBUTING.md
