### 2024-11-19 #7501 @paquiteau
Hi @Kevin-Mattheus-Moerman  thanks for starting the review process.

I have added the CONTRIBUTING.md and updated the bib references whenever possible (sadly ISMRM abstracts do not have DOIs, that is one of the reasons we are doing this JOSS submission actually)

For reviewers, I could suggest JeffFessler or guanhuaw (however, I am currently a reviewer for their JOSS submission for MIRTorch: https://github.com/openjournals/joss-reviews/issues/7340)

Then there is possibly uecker, mathieuboudreau, jonbmartin (also a reviewer of MIRTorch)
