### 2024-11-19 b9fe89d
fix: use asarray for torch compat.
