### 2024-11-21 66cd861
Fixes #211: Adjust normalization of density compensation coefficients. Add Cufinufft Pipe function authored by Chaithya G.R.
