+++ src/mrinufft/operators/interfaces/gpunufft.py
@@ -607,13 +607,9 @@ class MRIGpuNUFFT(FourierOperatorBase):
             max_iter=num_iterations
         )
         if normalize:
-            test_op = MRIGpuNUFFT(
-                samples=kspace_loc,
-                shape=original_shape,
-                **kwargs
-            )
+            test_op = MRIGpuNUFFT(samples=kspace_loc, shape=original_shape, **kwargs)
             test_im = np.ones(original_shape, dtype=np.complex64)
-            test_im_recon = test_op.adj_op(density_comp*test_op.op(test_im))
+            test_im_recon = test_op.adj_op(density_comp * test_op.op(test_im))
             density_comp /= np.mean(np.abs(test_im_recon))
         return density_comp.squeeze()
 
