### 2024-11-23 362ca9f
Add samplings to random walk initializations and fix boundaries
