### 2024-11-24 fdcda55
Add sampling density example skeleton
