+++ examples/example_sampling_densities.py
@@ -1,0 +1,387 @@
+"""
+==================
+Sampling densities
+==================
+
+A collection of sampling densities and density-based non-Cartesian trajectories.
+
+"""
+
+# %%
+# In this example we illustrate the use of different sampling densities,
+# and show how to generate trajectories based on them, such as random
+# walks and travelling-salesman trajectories.
+#
+
+# External
+import brainweb_dl as bwdl
+import matplotlib.pyplot as plt
+import numpy as np
+from utils import (
+    show_density,
+    show_densities,
+    show_locations,
+    show_trajectory,
+    show_trajectories,
+)
+
+# Internal
+import mrinufft as mn
+from mrinufft import display_2D_trajectory, display_3D_trajectory
+
+
+# %%
+# Script options
+# ==============
+#
+# These options are used in the examples below as default values for
+# all densities and trajectories.
+
+# Density parameters
+shape_2d = (100, 100)
+shape_3d = (100, 100, 100)
+
+# Trajectory parameters
+Nc = 10  # Number of shots
+Ns = 50  # Number of samples per shot
+
+# Display parameters
+figure_size = 5.5  # Figure size for trajectory plots
+subfigure_size = 3  # Figure size for subplots
+one_shot = 0  # Highlight one shot in particular
+
+# %%
+# Densities
+# =========
+#
+# In this section we present different sampling densities
+# with various properties.
+#
+# Cutoff/decay density
+# --------------------
+#
+# Create a density composed of a central constant-value ellipsis
+# defined by a cutoff ratio, followed by a polynomial decay over
+# outer regions as defined in [Cha+22]_.
+
+cutoff_density = mn.create_cutoff_decay_density(shape=shape_2d, cutoff=0.2, decay=2)
+show_density(cutoff_density, figure_size=figure_size)
+
+# %%
+# ``cutoff (float)``
+# ~~~~~~~~~~~~~~~~~~
+#
+# The k-space radius cutoff ratio between 0 and 1 within
+# which density remains uniform and beyond which it decays.
+# It is modulated by ``resolution`` to create ellipsoids.
+#
+# The ``mrinufft.trajectories.sampling.create_polynomial_density``
+# simply calls this function with ``cutoff=0``.
+
+arguments = [0, 0.1, 0.2, 0.3]
+function = lambda x: mn.create_cutoff_decay_density(
+    shape=shape_2d,
+    cutoff=x,
+    decay=2,
+)
+show_densities(
+    function,
+    arguments,
+    subfig_size=subfigure_size,
+)
+
+
+# %%
+# ``decay (float)``
+# ~~~~~~~~~~~~~~~~~
+#
+# The polynomial decay in density beyond the cutoff ratio.
+# It can be zero or negative as shown below, but most applications
+# are expected have decays in the positive range.
+
+
+arguments = [-1, 0, 0.5, 2]
+function = lambda x: mn.create_cutoff_decay_density(
+    shape=shape_2d,
+    cutoff=0.2,
+    decay=x,
+)
+show_densities(
+    function,
+    arguments,
+    subfig_size=subfigure_size,
+)
+
+# %%
+# ``resolution (tuple)``
+# ~~~~~~~~~~~~~~~~~~~~~~
+#
+# Resolution scaling factors for each dimension of the density grid,
+# by default ``None``. Note on the example below that the unit doesn't
+# matter because ``cutoff`` is a ratio and ``decay`` is an exponent,
+# so only the relative factor between the dimensions is important.
+#
[…]
