### 2024-11-25 96d9850
Apply suggestions from code review
Co-authored-by: Guillaume Daval-Frérot <gdavalfrerot@gmail.com>
