### 2024-11-25 2f631ca
Add descriptions to sampling density examples
