+++ examples/example_sampling_densities.py
@@ -75,7 +75,7 @@ show_density(cutoff_density, figure_size=figure_size)
 # which density remains uniform and beyond which it decays.
 # It is modulated by ``resolution`` to create ellipsoids.
 #
-# The ``mrinufft.trajectories.sampling.create_polynomial_density``
+# The ``mrinufft.create_polynomial_density``
 # simply calls this function with ``cutoff=0``.
 
 arguments = [0, 0.1, 0.2, 0.3]
@@ -90,7 +90,6 @@ show_densities(
     subfig_size=subfigure_size,
 )
 
-
 # %%
 # ``decay (float)``
 # ~~~~~~~~~~~~~~~~~
@@ -98,7 +97,6 @@ show_densities(
 # The polynomial decay in density beyond the cutoff ratio.
 # It can be zero or negative as shown below, but most applications
 # are expected have decays in the positive range.
-
 
 arguments = [-1, 0, 0.5, 2]
 function = lambda x: mn.create_cutoff_decay_density(
@@ -138,6 +136,7 @@ show_densities(
     subfig_size=subfigure_size,
 )
 
+
 # %%
 # Energy-based density
 # --------------------
@@ -164,7 +163,6 @@ show_density(energy_density, figure_size=figure_size, log_scale=True)
 # More relevant use cases would be to learn densities for
 # different organs and/or contrasts.
 
-
 arguments = [50, 100, 150]
 function = lambda x: mn.create_energy_density(dataset=bwdl.get_mri(4, "T1")[x : x + 20])
 show_densities(
@@ -179,7 +177,14 @@ show_densities(
 # Chauffert's density
 # -------------------
 #
-
+# This is a reproduction of the proposition from [CCW13]_.
+# A sampling density is derived from compressed sensing
+# equations to maximize guarantees of exact image recovery
+# for a specified sparse wavelet domain decomposition.
+#
+# This principle is valid for any linear transform but
+# for convenience it was limited to wavelets as in the
+# original implementation.
 
 chauffert_density = mn.create_chauffert_density(
     shape=shape_2d,
@@ -192,6 +197,10 @@ show_density(chauffert_density, figure_size=figure_size)
 # %%
 # ``wavelet_basis (str)``
 # ~~~~~~~~~~~~~~~~~~~~~~~
+#
+# The wavelet basis to use for wavelet decomposition, either
+# as a built-in wavelet name from the PyWavelets package
+# or as a custom ``pywt.Wavelet`` object.
 
 arguments = ["haar", "rbio2.2", "coif4", "sym8"]
 function = lambda x: mn.create_chauffert_density(
@@ -208,6 +217,8 @@ show_densities(
 # %%
 # ``nb_wavelet_scales (int)``
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#
+# The number of wavelet scales to use in decomposition.
 
 arguments = [1, 2, 3, 4]
 function = lambda x: mn.create_chauffert_density(
@@ -225,6 +236,9 @@ show_densities(
 # %%
 # Custom density
 # --------------
+#
+# Any density can be defined and later used for sampling and
+# trajectories.
 
 # Linear gradient
 density = np.tile(np.linspace(0, 1, shape_2d[1])[:, None], (1, shape_2d[0]))
@@ -269,6 +283,9 @@ show_densities(function, arguments, subfig_size=subfigure_size)
 # %%
 # Random sampling
 # ---------------
+#
+# This sampling simply consists of weighted-random selection from the
+# density grid locations.
 
 arguments = densities.keys()
 function = lambda x: mn.sample_from_density(Nc * Ns, densities[x], method="random")
@@ -278,6 +295,11 @@ show_locations(function, arguments, subfig_size=subfigure_size)
 # %%
 # Lloyd's sampling
 # ----------------
+#
[…]
