+++ examples/GPU/example_3d_trajectory_display.py
@@ -39,16 +39,14 @@ def plot_slices(axs, volume, title=""):
         set_labels(axs[i], i)
 
 
-def create_grid(grid_type, title="", **kwargs):
+def create_grid(grid_type, trajectories, **kwargs):
     fig, axs = plt.subplots(3, 3, figsize=(10, 10))
     for i, (name, traj) in enumerate(trajectories.items()):
         grid = get_gridded_trajectory(
-            traj, (64, 64, 64), grid_type=grid_type, traj_params=traj_params, **kwargs
+            traj, traj_params['img_size'], grid_type=grid_type, traj_params=traj_params, **kwargs
         )
         plot_slices(axs[:, i], grid, title=name)
     plt.tight_layout()
-    plt.suptitle(title)
-    plt.show()
 
 
 # %%
@@ -56,7 +54,7 @@ def create_grid(grid_type, title="", **kwargs):
 trajectories = {
     "Radial": mtt.initialize_3D_phyllotaxis_radial(64 * 8, 64),
     "FLORETs": mtt.initialize_3D_floret(64 * 8, 64, nb_revolutions=2),
-    "Yarn Ball": mtt.initialize_3D_seiffert_spiral(64 * 8, 64),
+    "Seiffert Spirals": mtt.initialize_3D_seiffert_spiral(64 * 8, 64),
 }
 traj_params = {
     "FOV": (0.23, 0.23, 0.23),
@@ -66,34 +64,44 @@ traj_params = {
 
 # %%
 # Display the density of the trajectories, along the 3 mid-planes. For this, make `grid_type="density"`.
-create_grid("density", "Sampling Density")
+create_grid("density", trajectories)
+plt.suptitle("Sampling Density")
+plt.show()
 
 
 # %%
-# Display the sampling times over the trajectories, along the 3 mid-planes. For this, make `grid_type="time"`.
+# Display the sampling times over the trajectories. For this, make `grid_type="time"`.
 # It helps to check the sampling times over the k-space trajectories, which can be responsible for excessive off-resonance artifacts.
-create_grid("time", "Sampling Time")
+create_grid("time", trajectories)
+plt.suptitle("Sampling Time")
+plt.show()
 
 # %%
-# Display the inversion time of the trajectories, along the 3 mid-planes. For this, make `grid_type="inversion"`.
+# Display the inversion time of the trajectories. For this, make `grid_type="inversion"`.
 # This helps in obtaining the inversion time when particular region of k-space is sampled, assuming the trajectories are time ordered.
 # This helps understand any issues for imaging involving inversion recovery.
 # The argument `turbo_factor` can be used to tell what is the number of echoes between 2 inversion pulses.
-create_grid("inversion", "Inversion Time", turbo_factor=64)
-
+create_grid("inversion", trajectories, turbo_factor=64)
+plt.suptitle("Inversion Time")
+plt.show()
 # %%
-# Display the k-space holes in the trajectories, along the 3 mid-planes. For this, make `grid_type="holes"`.
+# Display the k-space holes in the trajectories. For this, make `grid_type="holes"`.
 # This helps in understanding the k-space holes, and can help debug artifacts in reconstructions.
-create_grid("holes", "K-space Holes", threshold=1e-2)
-
+create_grid("holes", trajectories, threshold=1e-2)
+plt.suptitle("K-space Holes")
+plt.show()
 # %%
-# Display the gradient strength of the trajectories, along the 3 mid-planes. For this, make `grid_type="gradients"`.
+# Display the gradient strength of the trajectories. For this, make `grid_type="gradients"`.
 # This helps in understanding the gradient strength applied at specific k-space region.
[…]
