### 2024-11-28 83b9081
Add dependency to pywavelet
