### 2024-12-02 4cafcaa
refactor: use StrEnum.
