### 2024-12-02 c8591ea
fix: correct array shape.
