### 2024-12-02 82172be
Improve cutoff/decay density, fix dependencies
