### 2024-12-05 7a6f4a0
adding density normalization
