+++ src/mrinufft/operators/interfaces/cufinufft.py
@@ -906,6 +906,7 @@ class MRICufiNUFFT(FourierOperatorBase):
             raise ValueError(
                 "gpuNUFFT is not available, cannot " "estimate the density compensation"
             )
+        original_shape = volume_shape
         volume_shape = np.array([_next235beven(int(osf * i), 1) for i in volume_shape])
         grid_op = MRICufiNUFFT(
             samples=kspace_loc,
@@ -922,4 +923,9 @@ class MRICufiNUFFT(FourierOperatorBase):
                     grid_op.adj_op(density_comp.astype(grid_op.cpx_dtype))
                 ).squeeze()
             )
+        if normalize:
+            test_op = MRICufiNUFFT(samples=kspace_loc, shape=original_shape, **kwargs)
+            test_im = cp.ones(original_shape, dtype=test_op.cpx_dtype)
+            test_im_recon = test_op.adj_op(density_comp * test_op.op(test_im))
+            density_comp /= cp.mean(cp.abs(test_im_recon))
         return density_comp.squeeze()
