+++ src/mrinufft/operators/interfaces/cufinufft.py
@@ -18,6 +18,7 @@ from .utils import (
     pin_memory,
     sizeof_fmt,
 )
+from .utils.utils import _next235beven
 
 CUFINUFFT_AVAILABLE = CUPY_AVAILABLE
 try:
@@ -38,32 +39,6 @@ OPTS_FIELD_DECODE = {
 }
 
 DTYPE_R2C = {"float32": "complex64", "float64": "complex128"}
-
-
-def _next235beven(n, b):
-    """Find the next even integer not less than n.
-
-    This function finds the next even integer not less than n, with prime factors no
-    larger than 5, and is a multiple of b (where b is a number that only
-    has prime factors 2, 3, and 5).
-    It is used in particular with `pipe` density compensation estimation.
-    """
-    if n <= 2:
-        return 2
-    if n % 2 == 1:
-        n += 1  # make it even
-    nplus = n - 2  # to cancel out the +=2 at start of loop
-    numdiv = 2  # a dummy that is >1
-    while numdiv > 1 or nplus % b != 0:
-        nplus += 2  # stays even
-        numdiv = nplus
-        while numdiv % 2 == 0:
-            numdiv //= 2  # remove all factors of 2, 3, 5...
-        while numdiv % 3 == 0:
-            numdiv //= 3
-        while numdiv % 5 == 0:
-            numdiv //= 5
-    return nplus
 
 
 class RawCufinufftPlan:
@@ -904,7 +879,7 @@ class MRICufiNUFFT(FourierOperatorBase):
         """
         if CUFINUFFT_AVAILABLE is False:
             raise ValueError(
-                "gpuNUFFT is not available, cannot " "estimate the density compensation"
+                "cufinufft is not available, cannot estimate the density compensation"
             )
         original_shape = volume_shape
         volume_shape = np.array([_next235beven(int(osf * i), 1) for i in volume_shape])
