+++ _sources/generated/autoexamples/GPU/example_learn_samples_multicoil.rst
@@ -327,9 +327,9 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466063]/volatile/github-ci-mind-inria/gpu_runner/_work/_tool/Python/3.10.15/x64/lib/python3.10/site-packages/mrinufft/operators/autodiff.py:98: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
+      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466064]/volatile/github-ci-mind-inria/gpu_runner/_work/_tool/Python/3.10.15/x64/lib/python3.10/site-packages/mrinufft/operators/autodiff.py:98: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
       grad_traj = torch.transpose(torch.sum(grad_traj, dim=1), 0, 1).to(
-      1%|          | 1/100 [00:01<01:43,  1.04s/steps, loss=0.15466063]      1%|          | 1/100 [00:01<01:43,  1.04s/steps, loss=0.27207443]      2%|▏         | 2/100 [00:02<01:47,  1.09s/steps, loss=0.27207443]      2%|▏         | 2/100 [00:02<01:47,  1.09s/steps, loss=0.21599334]      3%|▎         | 3/100 [00:03<01:43,  1.07s/steps, loss=0.21599334]      3%|▎         | 3/100 [00:03<01:43,  1.07s/steps, loss=0.12370375]      4%|▍         | 4/100 [00:04<01:54,  1.19s/steps, loss=0.12370375]      4%|▍         | 4/100 [00:04<01:54,  1.19s/steps, loss=0.10386084]      5%|▌         | 5/100 [00:05<01:52,  1.18s/steps, loss=0.10386084]      5%|▌         | 5/100 [00:06<01:52,  1.18s/steps, loss=0.09464525]      6%|▌         | 6/100 [00:06<01:52,  1.19s/steps, loss=0.09464525]      6%|▌         | 6/100 [00:07<01:52,  1.19s/steps, loss=0.0787075]       7%|▋         | 7/100 [00:08<01:53,  1.23s/steps, loss=0.0787075]      7%|▋         | 7/100 [00:08<01:53,  1.23s/steps, loss=0.08207652]      8%|▊         | 8/100 [00:09<01:55,  1.25s/steps, loss=0.08207652]      8%|▊         | 8/100 [00:09<01:55,  1.25s/steps, loss=0.07630253]      9%|▉         | 9/100 [00:10<01:52,  1.24s/steps, loss=0.07630253]      9%|▉         | 9/100 [00:11<01:52,  1.24s/steps, loss=0.0674832]      10%|█         | 10/100 [00:12<01:52,  1.25s/steps, loss=0.0674832]     10%|█         | 10/100 [00:12<01:52,  1.25s/steps, loss=0.063778095]     11%|█         | 11/100 [00:13<01:48,  1.22s/steps, loss=0.063778095]     11%|█         | 11/100 [00:13<01:48,  1.22s/steps, loss=0.058604404]     12%|█▏        | 12/100 [00:14<01:42,  1.16s/steps, loss=0.058604404]     12%|█▏        | 12/100 [00:14<01:42,  1.16s/step
[…]
