+++ _sources/generated/autoexamples/example_2D_trajectories.rst
@@ -1338,7 +1338,7 @@ References
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 27.949 seconds)
+   **Total running time of the script:** (0 minutes 29.577 seconds)
 
 
 .. _sphx_glr_download_generated_autoexamples_example_2D_trajectories.py:
