+++ src/mrinufft/io/siemens.py
@@ -60,6 +60,12 @@ def read_siemens_rawdat(
             "The mapVBVD module is not available. Please install it using "
             "the following command: pip install pymapVBVD"
         ) from err
+    except ImportError as err:
+        raise ImportError(
+            "The mapVBVD module is not available. Please install "
+            "it along with the [extra] dependencies "
+            "or using `pip install pymapVBVD`."
+        ) from err
     twixObj = mapVBVD(filename)
     if isinstance(twixObj, list):
         twixObj = twixObj[-1]
