+++ pyproject.toml
@@ -17,9 +17,7 @@ cufinufft = ["cufinufft<2.3", "cupy-cuda12x"]
 finufft = ["finufft"]
 pynfft = ["pynfft2>=1.4.3; python_version < '3.12'", "numpy>=2.0.0; python_version < '3.12'"]
 pynufft = ["pynufft"]
-io = ["pymapvbvd"]
-smaps = ["scikit-image"]
-sampling = ["pywavelets", "scikit-learn"]
+extra = ["pymapvbvd", "scikit-image", "scikit-learn", "pywavelets"]
 autodiff = ["torch"]
 
 
