### 2024-12-10 432df36
Fix black format in smaps
