### 2024-12-10 4ff1bf0
Remove except duplicate in io/siemens
