+++ pyproject.toml
@@ -17,7 +17,7 @@ torchkbnufft = ["torchkbnufft", "cupy-cuda12x"]
 torchkbnufft-cpu = ["torchkbnufft", "cupy-cuda12x"]
 torchkbnufft-gpu = ["torchkbnufft", "cupy-cuda12x"]
 
-cufinufft = ["cufinufft>=2.3", "cupy-cuda12x"]
+cufinufft = ["cufinufft>=2.3.1", "cupy-cuda12x"]
 tensorflow = ["tensorflow-mri==0.21.0", "tensorflow-probability==0.17.0", "tensorflow-io==0.27.0", "matplotlib==3.7"]
 finufft = ["finufft>=2.3"]
 sigpy = ["sigpy"]
