### 2024-12-12 e60da32
Update CI with new [extra] dependencies
