+++ .github/workflows/test-ci.yml
@@ -206,7 +206,7 @@ jobs:
         shell: bash
         run: |
           python -m pip install --upgrade pip
-          python -m pip install -e .[test,dev]
+          python -m pip install -e .[extra,test,dev]
           python -m pip install finufft pooch brainweb-dl torch fastmri
         
       - name: Install GPU related interfaces
