### 2024-12-16 9c41feb
Densities, random walks & travelling salesman (#206)
* Split trajectory.maths.py into a submodule

* Add densities, random walk and travelling salesman trajectories

* fixup! Add densities, random walk and travelling salesman trajectories

* Add docstrings for random walk and TSP solver

* fixup! Add densities, random walk and travelling salesman trajectories

* Add docstrings to TSP & sampling densities

* Add samplings to random walk initializations and fix boundaries

* Add utils to examples to show densities

* Add sampling density example skeleton

* Add descriptions to sampling density e
[…]
