+++ src/mrinufft/trajectories/maths/rotations.py
@@ -1,0 +1,164 @@
+"""Rotation functions in 2D & 3D spaces."""
+
+import numpy as np
+import numpy.linalg as nl
+
+
+def R2D(theta):
+    """Initialize 2D rotation matrix.
+
+    Parameters
+    ----------
+    theta : float
+        Rotation angle in rad.
+
+    Returns
+    -------
+    np.ndarray
+        2D rotation matrix.
+    """
+    return np.array([[np.cos(theta), -np.sin(theta)], [np.sin(theta), np.cos(theta)]])
+
+
+def Rx(theta):
+    """Initialize 3D rotation matrix around x axis.
+
+    Parameters
+    ----------
+    theta : float
+        Rotation angle in rad.
+
+    Returns
+    -------
+    np.ndarray
+        3D rotation matrix.
+    """
+    return np.array(
+        [
+            [1, 0, 0],
+            [0, np.cos(theta), -np.sin(theta)],
+            [0, np.sin(theta), np.cos(theta)],
+        ]
+    )
+
+
+def Ry(theta):
+    """Initialize 3D rotation matrix around y axis.
+
+    Parameters
+    ----------
+    theta : float
+        Rotation angle in rad.
+
+    Returns
+    -------
+    np.ndarray
+        3D rotation matrix.
+    """
+    return np.array(
+        [
+            [np.cos(theta), 0, np.sin(theta)],
+            [0, 1, 0],
+            [-np.sin(theta), 0, np.cos(theta)],
+        ]
+    )
+
+
+def Rz(theta):
+    """Initialize 3D rotation matrix around z axis.
+
+    Parameters
+    ----------
+    theta : float
+        Rotation angle in rad.
+
+    Returns
+    -------
+    np.ndarray
+        3D rotation matrix.
+    """
+    return np.array(
+        [
+            [np.cos(theta), -np.sin(theta), 0],
+            [np.sin(theta), np.cos(theta), 0],
+            [0, 0, 1],
+        ]
+    )
+
+
+def Rv(v1, v2, normalize=True, eps=1e-8):
+    """Initialize 3D rotation matrix from two vectors.
+
+    Initialize a 3D rotation matrix from two vectors using Rodrigues's rotation
+    formula. Note that the rotation is carried around the axis orthogonal to both
+    vectors from the origin, and therefore is undetermined when both vectors
+    are colinear. While this case is handled manually, close cases might result
+    in approximative behaviors.
+
+    Parameters
+    ----------
+    v1 : np.ndarray
+        Source vector.
+    v2 : np.ndarray
+        Target vector.
+    normalize : bool, optional
+        Normalize the vectors. The default is True.
+
+    Returns
+    -------
+    np.ndarray
+        3D rotation matrix.
+    """
+    # Check for colinearity, not handled by Rodrigues' coefficients
+    if nl.norm(np.cross(v1, v2)) < eps:
+        sign = np.sign(np.dot(v1, v2))
+        return sign * np.identity(3)
+
+    if normalize:
+        v1, v2 = v1 / np.linalg.norm(v1), v2 / np.linalg.norm(v2)
+    cos_theta = np.dot(v1, v2)
+    v3 = np.cross(v1, v2)
+    cross_matrix = np.cross(v3, np.identity(v3.shape[0]) * -1)
+    return np.identity(3) + cross_matrix + cross_matrix @ cross_matrix / (1 + cos_theta)
+
+
+def Ra(vector, theta):
+    """Initialize 3D rotation matrix around an arbitrary vector.
+
+    Initialize a 3D rotation matrix to rotate around `vector` by an angle `theta`.
[…]
