+++ _sources/generated/_autosummary/mrinufft.trajectories.inits.rst
@@ -1,0 +1,30 @@
+``trajectories.inits``
+======================
+
+.. automodule:: mrinufft.trajectories.inits
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
+.. autosummary::
+   :toctree:
+   :recursive:
+
+   random_walk
+   travelling_salesman
+
