+++ _sources/generated/_autosummary/mrinufft.trajectories.sampling.create_energy_density.rst
@@ -1,0 +1,11 @@
+create\_energy\_density
+=======================
+
+.. currentmodule:: mrinufft.trajectories.sampling
+
+.. autofunction:: create_energy_density
+
+..  _sphx_glr_backref_mrinufft.trajectories.sampling.create_energy_density:
+
+.. minigallery:: mrinufft.trajectories.sampling.create_energy_density
+   :add-heading:
