+++ _sources/generated/_autosummary/mrinufft.trajectories.maths.rotations.Rx.rst
@@ -1,0 +1,11 @@
+Rx
+==
+
+.. currentmodule:: mrinufft.trajectories.maths.rotations
+
+.. autofunction:: Rx
+
+..  _sphx_glr_backref_mrinufft.trajectories.maths.rotations.Rx:
+
+.. minigallery:: mrinufft.trajectories.maths.rotations.Rx
+   :add-heading:
