+++ _sources/generated/_autosummary/mrinufft.trajectories.maths.get_closest_fibonacci_number.rst
@@ -1,11 +1,0 @@
-get\_closest\_fibonacci\_number
-===============================
-
-.. currentmodule:: mrinufft.trajectories.maths
-
-.. autofunction:: get_closest_fibonacci_number
-
-..  _sphx_glr_backref_mrinufft.trajectories.maths.get_closest_fibonacci_number:
-
-.. minigallery:: mrinufft.trajectories.maths.get_closest_fibonacci_number
-   :add-heading:
