+++ _sources/generated/autoexamples/example_2D_trajectories.rst
@@ -40,7 +40,7 @@ will refer to previous patterns for explanation.
     # External
     import matplotlib.pyplot as plt
     import numpy as np
-    from utils import show_argument, show_trajectory
+    from utils import show_trajectories, show_trajectory
 
     # Internal
     import mrinufft as mn
@@ -142,7 +142,7 @@ k-space. More shots means better coverage but also longer acquisitions.
 
     arguments = [8, 16, 32, 64]
     function = lambda x: mn.initialize_2D_radial(x, Ns, tilt=tilt, in_out=in_out)
-    show_argument(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+    show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
 
 
 
@@ -174,7 +174,7 @@ is lengthened or the sampling rate is increased.
 
     arguments = [8, 16, 32, 64]
     function = lambda x: mn.initialize_2D_radial(Nc, x, tilt=tilt, in_out=in_out)
-    show_argument(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+    show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
 
 
 
@@ -208,7 +208,7 @@ when the ``in_out`` argument is switched to keep the same behavior.
 
     arguments = ["uniform", "golden", "mri-golden", np.pi / 17]
     function = lambda x: mn.initialize_2D_radial(Nc, Ns, tilt=x, in_out=in_out)
-    show_argument(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+    show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
 
 
 
@@ -249,7 +249,7 @@ to update it too when switching ``in_out``.
 
     arguments = [True, False]
     function = lambda x: mn.initialize_2D_radial(Nc, Ns, tilt=tilt, in_out=x)
-    show_argument(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+    show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
 
 
 
@@ -327,7 +327,7 @@ twice for in-out trajectories).
     function = lambda x: mn.initialize_2D_spiral(
         Nc, Ns, tilt=tilt, nb_revolutions=x, in_out=in_out
     )
-    show_argument(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+    show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
 
 
 
@@ -369,7 +369,7 @@ of their asymptotic behavior around the center.
 
     arguments = ["galilean", "archimedes", "fermat", 1 / 4]
     function = lambda x: mn.initialize_2D_spiral(Nc, Ns, tilt=tilt, spiral=x, in_out=in_out)
-    show_argument(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+    show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
 
 
 
@@ -414,7 +414,7 @@ subtle, as shown below.
         Ns,
         patch_center=x,
     )
-    show_argument(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+    show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
 
 
 
@@ -496,7 +496,7 @@ spiral length, which makes k-space denser along the shorter shots.
         Ns,
         spiral_reduction=x,
     )
[…]
