+++ _sources/generated/autoexamples/GPU/example_learn_straight_line_readouts.rst
@@ -330,7 +330,7 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.31016502]      1%|          | 1/100 [00:05<08:50,  5.36s/steps, loss=0.31016502]      1%|          | 1/100 [00:05<08:50,  5.36s/steps, loss=0.2797733]       2%|▏         | 2/100 [00:10<08:13,  5.04s/steps, loss=0.2797733]      2%|▏         | 2/100 [00:10<08:13,  5.04s/steps, loss=0.2674313]      3%|▎         | 3/100 [00:14<07:49,  4.84s/steps, loss=0.2674313]      3%|▎         | 3/100 [00:14<07:49,  4.84s/steps, loss=0.26029155]      4%|▍         | 4/100 [00:20<08:00,  5.00s/steps, loss=0.26029155]      4%|▍         | 4/100 [00:20<08:00,  5.00s/steps, loss=0.25727516]      5%|▌         | 5/100 [00:25<08:09,  5.15s/steps, loss=0.25727516]      5%|▌         | 5/100 [00:25<08:09,  5.15s/steps, loss=0.2528005]       6%|▌         | 6/100 [00:30<08:10,  5.22s/steps, loss=0.2528005]      6%|▌         | 6/100 [00:30<08:10,  5.22s/steps, loss=0.25024423]      7%|▋         | 7/100 [00:36<08:25,  5.43s/steps, loss=0.25024423]      7%|▋         | 7/100 [00:36<08:25,  5.43s/steps, loss=0.24693525]      8%|▊         | 8/100 [00:42<08:19,  5.43s/steps, loss=0.24693525]      8%|▊         | 8/100 [00:42<08:19,  5.43s/steps, loss=0.24176724]      9%|▉         | 9/100 [00:47<08:21,  5.52s/steps, loss=0.24176724]      9%|▉         | 9/100 [00:48<08:21,  5.52s/steps, loss=0.23845014]     10%|█         | 10/100 [00:52<08:04,  5.39s/steps, loss=0.23845014]     10%|█         | 10/100 [00:53<08:04,  5.39s/steps, loss=0.23644473]     11%|█         | 11/100 [00:57<07:50,  5.29s/steps, loss=0.23644473]     11%|█         | 11/100 [00:58<07:50,  5.29s/steps, loss=0.23593816]     12%|█▏        | 12/100 [01:03<08:01,  5.47s/steps, loss=0.23593816]     12%|█▏        | 12/100 [01:04<08:01,  5.47s/steps, loss=0.23418666]     13%|█▎        | 13/100 [01:08<07:37,  5.26s/steps, loss=0.23418666]     13%|█▎        | 13/100 [01:08<07:37,  5.26s/steps, loss=0.23275968]     14%|█▍        | 14/100 [01:13<07:22,  5.15s/steps, loss=0.23275968]     14%|█▍        | 14/100 [01:13<07:22,  5.15s/steps, loss=0.23295936]     15%|█▌        | 15/100 [01:18<07:17,  5.15s/steps, loss=0.23295936]     15%|█▌        | 15/100 [01:18<07:17,  5.15s/steps, loss=0.23299333]     16%|█▌        | 16/100 [01:23<07:00,  5.01s/steps, loss=0.23299333]     16%|█▌        | 16/100 [01:23<07:00,  5.01s/steps, loss=0.2320924]      17%|█▋        | 17/100 [01:28<06:55,  5.00s/steps, loss=0.2320924]     17%|█▋      
[…]
