+++ _sources/generated/_autosummary/mrinufft.trajectories.maths.fibonacci.rst
@@ -1,0 +1,35 @@
+``maths.fibonacci``
+===================
+
+.. automodule:: mrinufft.trajectories.maths.fibonacci
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree:
+      :nosignatures:
+   
+      generate_fibonacci_circle
+      generate_fibonacci_lattice
+      generate_fibonacci_sphere
+      get_closest_fibonacci_number
+      is_from_fibonacci_sequence
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
