+++ _sources/generated/_autosummary/mrinufft.trajectories.sampling.create_chauffert_density.rst
@@ -1,0 +1,11 @@
+create\_chauffert\_density
+==========================
+
+.. currentmodule:: mrinufft.trajectories.sampling
+
+.. autofunction:: create_chauffert_density
+
+..  _sphx_glr_backref_mrinufft.trajectories.sampling.create_chauffert_density:
+
+.. minigallery:: mrinufft.trajectories.sampling.create_chauffert_density
+   :add-heading:

[…]
