### 2024-12-18 b4fd267
feat: add example
