### 2024-12-18 d526791
Clean documentation and add type hints
