+++ src/mrinufft/trajectories/trajectory3D.py
@@ -24,7 +24,9 @@ from .utils import KMAX, Packings, initialize_shape_norm, initialize_tilt
 ##############
 
 
-def initialize_3D_phyllotaxis_radial(Nc, Ns, nb_interleaves=1, in_out=False):
+def initialize_3D_phyllotaxis_radial(
+    Nc: int, Ns: int, nb_interleaves: int = 1, in_out: bool = False
+) -> np.ndarray:
     """Initialize 3D radial trajectories with phyllotactic structure.
 
     The radial shots are oriented according to a Fibonacci sphere
@@ -53,7 +55,7 @@ def initialize_3D_phyllotaxis_radial(Nc, Ns, nb_interleaves=1, in_out=False):
 
     Returns
     -------
-    array_like
+    np.ndarray
         3D phyllotaxis radial trajectory
 
     References
@@ -71,7 +73,9 @@ def initialize_3D_phyllotaxis_radial(Nc, Ns, nb_interleaves=1, in_out=False):
     return trajectory
 
 
-def initialize_3D_golden_means_radial(Nc, Ns, in_out=False):
+def initialize_3D_golden_means_radial(
+    Nc: int, Ns: int, in_out: bool = False
+) -> np.ndarray:
     """Initialize 3D radial trajectories with golden means-based structure.
 
     The radial shots are oriented using multidimensional golden means,
@@ -95,7 +99,7 @@ def initialize_3D_golden_means_radial(Nc, Ns, in_out=False):
 
     Returns
     -------
-    array_like
+    np.ndarray
         3D golden means radial trajectory
 
     References
@@ -123,7 +127,9 @@ def initialize_3D_golden_means_radial(Nc, Ns, in_out=False):
     return KMAX * trajectory
 
 
-def initialize_3D_wong_radial(Nc, Ns, nb_interleaves=1, in_out=False):
+def initialize_3D_wong_radial(
+    Nc: int, Ns: int, nb_interleaves: int = 1, in_out: bool = False
+) -> np.ndarray:
     """Initialize 3D radial trajectories with a spiral structure.
 
     The radial shots are oriented according to an archimedean spiral
@@ -149,7 +155,7 @@ def initialize_3D_wong_radial(Nc, Ns, nb_interleaves=1, in_out=False):
 
     Returns
     -------
-    array_like
+    np.ndarray
         3D Wong radial trajectory
 
     References
@@ -181,7 +187,9 @@ def initialize_3D_wong_radial(Nc, Ns, nb_interleaves=1, in_out=False):
     return trajectory
 
 
-def initialize_3D_park_radial(Nc, Ns, nb_interleaves=1, in_out=False):
+def initialize_3D_park_radial(
+    Nc: int, Ns: int, nb_interleaves: int = 1, in_out: bool = False
+) -> np.ndarray:
     """Initialize 3D radial trajectories with a spiral structure.
 
     The radial shots are oriented according to an archimedean spiral
@@ -208,7 +216,7 @@ def initialize_3D_park_radial(Nc, Ns, nb_interleaves=1, in_out=False):
 
     Returns
     -------
-    array_like
+    np.ndarray
         3D Park radial trajectory
 
     References
@@ -233,8 +241,14 @@ def initialize_3D_park_radial(Nc, Ns, nb_interleaves=1, in_out=False):
 
 
 def initialize_3D_cones(
-    Nc, Ns, tilt="golden", in_out=False, nb_zigzags=5, spiral="archimedes", width=1
-):
+    Nc: int,
+    Ns: int,
+    tilt: str | float = "golden",
+    in_out: bool = False,
[…]
