+++ src/mrinufft/trajectories/maths/rotations.py
@@ -4,7 +4,7 @@ import numpy as np
 import numpy.linalg as nl
 
 
-def R2D(theta):
+def R2D(theta: float) -> np.ndarray:
     """Initialize 2D rotation matrix.
 
     Parameters
@@ -20,7 +20,7 @@ def R2D(theta):
     return np.array([[np.cos(theta), -np.sin(theta)], [np.sin(theta), np.cos(theta)]])
 
 
-def Rx(theta):
+def Rx(theta: float) -> np.ndarray:
     """Initialize 3D rotation matrix around x axis.
 
     Parameters
@@ -42,7 +42,7 @@ def Rx(theta):
     )
 
 
-def Ry(theta):
+def Ry(theta: float) -> np.ndarray:
     """Initialize 3D rotation matrix around y axis.
 
     Parameters
@@ -64,7 +64,7 @@ def Ry(theta):
     )
 
 
-def Rz(theta):
+def Rz(theta: float) -> np.ndarray:
     """Initialize 3D rotation matrix around z axis.
 
     Parameters
@@ -86,7 +86,9 @@ def Rz(theta):
     )
 
 
-def Rv(v1, v2, normalize=True, eps=1e-8):
+def Rv(
+    v1: np.ndarray, v2: np.ndarray, eps: float = 1e-8, *, normalize: bool = True
+) -> np.ndarray:
     """Initialize 3D rotation matrix from two vectors.
 
     Initialize a 3D rotation matrix from two vectors using Rodrigues's rotation
@@ -101,6 +103,8 @@ def Rv(v1, v2, normalize=True, eps=1e-8):
         Source vector.
     v2 : np.ndarray
         Target vector.
+    eps : float, optional
+        Tolerance to consider two vectors as colinear. The default is 1e-8.
     normalize : bool, optional
         Normalize the vectors. The default is True.
 
@@ -122,7 +126,7 @@ def Rv(v1, v2, normalize=True, eps=1e-8):
     return np.identity(3) + cross_matrix + cross_matrix @ cross_matrix / (1 + cos_theta)
 
 
-def Ra(vector, theta):
+def Ra(vector: np.ndarray, theta: float) -> np.ndarray:
     """Initialize 3D rotation matrix around an arbitrary vector.
 
     Initialize a 3D rotation matrix to rotate around `vector` by an angle `theta`.

[…]
