### 2024-12-19 630486c
feat: update example and tweaks
