### 2024-12-19 2ece4be
feat: update example and tweaks
