+++ examples/example_trajectory_tools.py
@@ -296,9 +296,10 @@ trajectory = tools.stack_random(
     planar_trajectories["Spiral"],
     dim_size=128,
     center_prop=0.1,
-    acceleration=4,
+    accel=16,
     pdf="uniform",
-    order="center-out",
+    order="top-down",
+    rng=42,
 )
 
 show_trajectory(trajectory, figure_size=figure_size, one_shot=one_shot)
@@ -308,9 +309,125 @@ show_trajectory(trajectory, figure_size=figure_size, one_shot=one_shot)
 # ~~~~~~~~~~~~~~~~~~~~~~
 # The main use case is to stack trajectories consisting of
 # flat or thick planes that will match the image slices.
-#
-#
-#
+arguments = ["Radial", "Spiral", "2D Cones", "3D Cones"]
+function = lambda x: tools.stack_random(
+    planar_trajectories[x],
+    dim_size=128,
+    center_prop=0.1,
+    accel=16,
+    pdf="gaussian",
+    order="top-down",
+    rng=42,
+)
+show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+
+# %%
+# ``dim_size (int)``
+# ~~~~~~~~~~~~~~~~~~
+# Size of the k-space in voxel units. It
+# is used to normalize the stack positions, and is used with the ``accel``
+# factor and ``center_prop`` to determine the number of stacks.
+arguments = [16, 32, 64, 128]
+function = lambda x: tools.stack_random(
+    planar_trajectories["Spiral"],
+    dim_size=x,
+    center_prop=0.1,
+    accel=16,
+    pdf="gaussian",
+    order="top-down",
+    rng=42,
+)
+show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+
+# %%
+# ``center_prop (int or float)``
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# Number of lines to keep in the center of the k-space. It is used to determine
+# the number of stacks and the acceleration factor. It is typically used to keep
+# the center of the k-space with a higher density of shots.
+# If a ``float`` this is a fraction of the total dim_size. If ``int`` it is
+# directly the number of lines.
+
+arguments = [1, 5, 0.1, 0.5]
+function = lambda x: tools.stack_random(
+    planar_trajectories["Spiral"],
+    dim_size=128,
+    center_prop=x,
+    accel=16,
+    pdf="uniform",
+    order="top-down",
+    rng=42,
+)
+show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+
+
+# %%
+# ``accel (int)``
+# ~~~~~~~~~~~~~~~
+# Acceleration factor to subsamples the outer region of the k-space.
+# Note that the acceleration factor does not take into account the center lines.
+
+
+arguments = [1, 4, 8, 16, 32]
+function = lambda x: tools.stack_random(
+    planar_trajectories["Spiral"],
+    dim_size=128,
+    center_prop=0.1,
+    accel=x,
+    pdf="uniform",
+    order="top-down",
+    rng=42,
+)
+show_trajectories(function, arguments, one_shot=one_shot, subfig_size=subfigure_size)
+
+# %%
+# ``pdf (str or array)``
+# ~~~~~~~~~~~~~~~~~~~~~~
+# Probability density function for the sampling of the outer region. It can
+# either be a string to use a known probability law ("gaussian" or "uniform") or
+# "equispaced" for a coherent undersampling (like the one used in GRAPPA). It
+# can also be a array, for using a customed density probability.
[…]
