### 2024-12-19 a0031a3
Apply ruff ANN rules to trajectories
