+++ src/mrinufft/trajectories/trajectory3D.py
@@ -1,6 +1,7 @@
 """Functions to initialize 3D trajectories."""
 
 from functools import partial
+from typing import Literal
 
 import numpy as np
 import numpy.linalg as nl
@@ -243,10 +244,10 @@ def initialize_3D_park_radial(
 def initialize_3D_cones(
     Nc: int,
     Ns: int,
-    tilt: str | float = "golden",
+    tilt: Literal | float = "golden",
     in_out: bool = False,
     nb_zigzags: float = 5,
-    spiral: str | float = "archimedes",
+    spiral: Literal | float = "archimedes",
     width: float = 1,
 ) -> np.ndarray:
     """Initialize 3D trajectories with cones.
@@ -264,14 +265,14 @@ def initialize_3D_cones(
         Number of shots
     Ns : int
         Number of samples per shot
-    tilt : str, float, optional
+    tilt : Literal, float, optional
         Tilt of the cones, by default "golden"
     in_out : bool, optional
         Whether the curves are going in-and-out or start from the center,
         by default False
     nb_zigzags : float, optional
         Number of zigzags of the cones, by default 5
-    spiral : str, float, optional
+    spiral : Literal, float, optional
         Spiral type, by default "archimedes"
     width : float, optional
         Cone width normalized such that `width=1` avoids cone overlaps, by default 1
@@ -326,8 +327,8 @@ def initialize_3D_floret(
     Ns: int,
     in_out: bool = False,
     nb_revolutions: float = 1,
-    spiral: str | float = "fermat",
-    cone_tilt: str | float = "golden",
+    spiral: Literal | float = "fermat",
+    cone_tilt: Literal | float = "golden",
     max_angle: float = np.pi / 2,
     axes: tuple[int, ...] = (2,),
 ) -> np.ndarray:
@@ -348,9 +349,9 @@ def initialize_3D_floret(
         Whether to start from the center or not, by default False
     nb_revolutions : float, optional
         Number of revolutions of the spirals, by default 1
-    spiral : str, float, optional
+    spiral : Literal, float, optional
         Spiral type, by default "fermat"
-    cone_tilt : str, float, optional
+    cone_tilt : Literal, float, optional
         Tilt of the cones around the :math:`k_z`-axis, by default "golden"
     max_angle : float, optional
         Maximum polar angle starting from the :math:`k_x-k_y` plane,
@@ -405,8 +406,8 @@ def initialize_3D_wave_caipi(
     Ns: int,
     nb_revolutions: float = 5,
     width: float = 1,
-    packing: str = "triangular",
-    shape: str | float = "square",
+    packing: Literal = "triangular",
+    shape: Literal | float = "square",
     spacing: tuple[int, int] = (1, 1),
 ) -> np.ndarray:
     """Initialize 3D trajectories with Wave-CAIPI.
@@ -425,11 +426,11 @@ def initialize_3D_wave_caipi(
         Diameter of the helices normalized such that `width=1`
         densely covers the k-space without overlap for square packing,
         by default 1.
-    packing : str, optional
+    packing : Literal, optional
         Packing method used to position the helices:
         "triangular"/"hexagonal", "square", "circular"
         or "random"/"uniform", by default "triangular".
-    shape : str or float, optional
+    shape : Literal or float, optional
         Shape over the 2D :math:`k_x-k_y` plane to pack with shots,
         either defined as `str` ("circle", "square", "diamond")
         or as `float` through p-norms following the conventions
@@ -524,8 +525,8 @@ def initialize_3D_seiffert_spiral(
     Ns: int,
     curve_index: float = 0.2,
     nb_revolutions: float = 1,
-    axis_tilt: str | float = "golden",
[…]

[…]
