### 2024-12-19 d5b3e09
Replace np.ndarray type hints with np.typing.NDArray
