### 2024-12-20 906f363
Prevent circular import
