+++ src/mrinufft/operators/interfaces/gpunufft.py
@@ -614,7 +614,7 @@ class MRIGpuNUFFT(FourierOperatorBase):
                 shape=original_shape,
                 density=density_comp,
                 backend=cls.backend,
-                **kwargs
+                **kwargs,
             )
         return density_comp.squeeze()
 
