### 2024-12-20 ba53be9
Some fixes and updates
