### 2024-12-25 914fc3f
Fix Literal use, change types to NDArray and reduce mypy errors
