+++ src/mrinufft/trajectories/tools.py
@@ -3,6 +3,7 @@
 from typing import Any, Callable, Literal
 
 import numpy as np
+from numpy.typing import NDArray
 from scipy.interpolate import CubicSpline, interp1d
 
 from .maths import Rv, Rx, Ry, Rz
@@ -14,28 +15,28 @@ from .utils import KMAX, initialize_tilt
 
 
 def stack(
-    trajectory: np.typing.NDArray,
+    trajectory: NDArray,
     nb_stacks: int,
-    z_tilt: Literal | float | None = None,
+    z_tilt: str | float | None = None,
     *,
     hard_bounded: bool = True,
-) -> np.ndarray:
+) -> NDArray:
     """Stack 2D or 3D trajectories over the :math:`k_z`-axis.
 
     Parameters
     ----------
-    trajectory : np.ndarray
+    trajectory : NDArray
         Trajectory in 2D or 3D to stack.
     nb_stacks : int
         Number of stacks repeating the provided trajectory.
-    z_tilt : Literal, float, optional
+    z_tilt : str | float, optional
         Tilt of the stacks, by default `None`.
     hard_bounded : bool, optional
         Whether the stacks should be strictly within the limits of the k-space.
 
     Returns
     -------
-    np.ndarray
+    NDArray
         Stacked trajectory.
     """
     # Check dimensionality and initialize output
@@ -64,30 +65,30 @@ def stack(
 
 
 def rotate(
-    trajectory: np.typing.NDArray,
+    trajectory: NDArray,
     nb_rotations: int,
-    x_tilt: Literal | float | None = None,
-    y_tilt: Literal | float | None = None,
-    z_tilt: Literal | float | None = None,
-) -> np.ndarray:
+    x_tilt: str | float | None = None,
+    y_tilt: str | float | None = None,
+    z_tilt: str | float | None = None,
+) -> NDArray:
     """Rotate 2D or 3D trajectories over the different axes.
 
     Parameters
     ----------
-    trajectory : np.ndarray
+    trajectory : NDArray
         Trajectory in 2D or 3D to rotate.
     nb_rotations : int
         Number of rotations repeating the provided trajectory.
-    x_tilt : Literal, optional
+    x_tilt : str | float, optional
         Tilt of the trajectory over the :math:`k_x`-axis, by default `None`.
-    y_tilt : Literal, optional
+    y_tilt : str | float, optional
         Tilt of the trajectory over the :math:`k_y`-axis, by default `None`.
-    z_tilt : Literal, optional
+    z_tilt : str | float, optional
         Tilt of the trajectory over the :math:`k_z`-axis, by default `None`.
 
     Returns
     -------
-    np.ndarray
+    NDArray
         Rotated trajectory.
     """
     # Check dimensionality and initialize output
@@ -110,33 +111,33 @@ def rotate(
 
 
 def precess(
-    trajectory: np.typing.NDArray,
+    trajectory: NDArray,
     nb_rotations: int,
-    tilt: Literal | float = "golden",
+    tilt: str | float = "golden",
     half_sphere: bool = False,
-    partition: Literal = "axial",
-    axis: int | np.ndarray | None = None,
-) -> np.ndarray:
+    partition: Literal["axial", "polar"] = "axial",
+    axis: int | NDArray | None = None,
+) -> NDArray:
     """Rotate trajectories as a precession around the :math:`k_z`-axis.
 
     Parameters
     ----------
-    trajectory : np.ndarray
+    trajectory : NDArray
         Trajectory in 2D or 3D to rotate.
     nb_rotations : int
         Number of rotations repeating the provided trajectory while precessing.
-    tilt : Literal, float, optional
+    tilt : str | float, optional
[…]
