### 2024-12-27 ff493e9
Add type alias for clustering coordinates
