+++ .github/workflows/test-ci.yml
@@ -183,7 +183,7 @@ jobs:
         shell: bash
         run: |
           python -m pip install --upgrade pip
-          python -m pip install -e --no-binary cufinufft .[test,dev,extra,finufft,cufinufft,gpuNUFFT,sigpy,autodiff,doc]
+          python -m pip install --no-binary cufinufft -e .[test,dev,extra,finufft,cufinufft,gpuNUFFT,sigpy,autodiff,doc]
         
       - name: Run examples
         shell: bash
