### 2025-01-09 702eb4e
proj: update numpy and minimal python version
Python 3.8 is unsupported since 2024-10. it's time to move on.
