+++ pyproject.toml
@@ -4,8 +4,8 @@ description = "MRI Non-Cartesian Fourier Operators with multiple computation bac
 authors = [{name="Pierre-antoine Comby", email="pierre-antoine.comby@crans.org"}]
 
 readme = "README.md"
-dependencies = ["numpy<=2.0.0", "scipy", "matplotlib", "tqdm"]
-requires-python = ">=3.8"
+dependencies = ["numpy", "scipy", "matplotlib", "tqdm"]
+requires-python = ">=3.9"
 
 dynamic = ["version"]
 
