+++ _sources/generated/autoexamples/example_readme.rst
@@ -64,7 +64,7 @@ An example to show how to perform a simple NUFFT.
 
  .. code-block:: none
 
-    /volatile/github-ci-mind-inria/gpu_runner/_work/_tool/Python/3.10.15/x64/lib/python3.10/site-packages/mrinufft/_utils.py:94: UserWarning: Samples will be rescaled to [-pi, pi), assuming they were in [-0.5, 0.5)
+    /volatile/github-ci-mind-inria/gpu_runner2/_work/_tool/Python/3.10.16/x64/lib/python3.10/site-packages/mrinufft/_utils.py:94: UserWarning: Samples will be rescaled to [-pi, pi), assuming they were in [-0.5, 0.5)
       warnings.warn(
 
 
@@ -123,7 +123,7 @@ An example to show how to perform a simple NUFFT.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 2.578 seconds)
+   **Total running time of the script:** (0 minutes 2.105 seconds)
 
 
 .. _sphx_glr_download_generated_autoexamples_example_readme.py:
