+++ _sources/generated/autoexamples/example_sampling_densities.rst
@@ -856,7 +856,7 @@ References
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (2 minutes 10.999 seconds)
+   **Total running time of the script:** (2 minutes 23.769 seconds)
 
 
 .. _sphx_glr_download_generated_autoexamples_example_sampling_densities.py:
