+++ _sources/generated/autoexamples/GPU/example_learn_samples_multicoil.rst
@@ -334,7 +334,9 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15462142]      1%|          | 1/100 [00:00<01:14,  1.33steps/s, loss=0.15462142]      1%|          | 1/100 [00:00<01:14,  1.33steps/s, loss=0.2685126]       2%|▏         | 2/100 [00:01<01:12,  1.34steps/s, loss=0.2685126]      2%|▏         | 2/100 [00:01<01:12,  1.34steps/s, loss=0.21949866]      3%|▎         | 3/100 [00:02<01:11,  1.35steps/s, loss=0.21949866]      3%|▎         | 3/100 [00:02<01:11,  1.35steps/s, loss=0.12228894]      4%|▍         | 4/100 [00:03<01:12,  1.33steps/s, loss=0.12228894]      4%|▍         | 4/100 [00:03<01:12,  1.33steps/s, loss=0.11065939]      5%|▌         | 5/100 [00:03<01:12,  1.30steps/s, loss=0.11065939]      5%|▌         | 5/100 [00:03<01:12,  1.30steps/s, loss=0.10674719]      6%|▌         | 6/100 [00:04<01:11,  1.31steps/s, loss=0.10674719]      6%|▌         | 6/100 [00:04<01:11,  1.31steps/s, loss=0.083452284]      7%|▋         | 7/100 [00:05<01:09,  1.33steps/s, loss=0.083452284]      7%|▋         | 7/100 [00:05<01:09,  1.33steps/s, loss=0.08412791]       8%|▊         | 8/100 [00:06<01:15,  1.21steps/s, loss=0.08412791]      8%|▊         | 8/100 [00:06<01:15,  1.21steps/s, loss=0.079916626]      9%|▉         | 9/100 [00:06<01:12,  1.26steps/s, loss=0.079916626]      9%|▉         | 9/100 [00:07<01:12,  1.26steps/s, loss=0.06711985]      10%|█         | 10/100 [00:07<01:09,  1.29steps/s, loss=0.06711985]     10%|█         | 10/100 [00:07<01:09,  1.29steps/s, loss=0.061132967]     11%|█         | 11/100 [00:08<01:06,  1.33steps/s, loss=0.061132967]     11%|█         | 11/100 [00:08<01:06,  1.33steps/s, loss=0.06320516]      12%|█▏        | 12/100 [00:09<01:05,  1.34steps/s, loss=0.06320516]     12%|█▏        | 12/100 [00:09<01:05,  1.34steps/s, loss=0.061265394]     13%|█▎        | 13/100 [00:09<01:04,  1.35steps/s, loss=0.061265394]     13%|█▎        | 13/100 [00:10<01:04,  1.35steps/s, loss=0.05457127]      14%|█▍        | 14/100 [00:10<01:02,  1.37steps/s, loss=0.05457127]     14%|█▍        | 14/100 [00:10<01:02,  1.37steps/s, loss=0.057657465]     15%|█▌        | 15/100 [00:11<01:02,  1.37steps/s, loss=0.057657465]     15%|█▌        | 15/100 [00:11<01:02,  1.37steps/s, loss=0.05254972]      16%|█▌        | 16/100 [00:12<01:01,  1.37steps/s, loss=0.05254972]     16%|█▌        | 16/100 [00:12<01:01,  1.37steps/s, loss=0.04896649]     17%|█▋        | 17/100 [00:12<01:01,  1.36steps/s, loss=0.04896649]     17%|█▋        | 17/100 [00:12<
[…]
