### 2025-01-10 3abd8d9
Clean documentation and add type hints (#228)
* Clean documentation and add type hints

* Apply ruff ANN rules to trajectories

* Replace np.ndarray type hints with np.typing.NDArray

* Fix missing import

* Fix Literal use, change types to NDArray and reduce mypy errors

* Add type alias for clustering coordinates
