+++ src/mrinufft/trajectories/sampling.py
@@ -1,17 +1,27 @@
 """Sampling densities and methods."""
 
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Literal
+
+if TYPE_CHECKING:
+    import pywt as pw
+
 import numpy as np
 import numpy.fft as nf
-import numpy.linalg as nl
-import numpy.random as nr
+from numpy.typing import NDArray
 from tqdm.auto import tqdm
 
 from .utils import KMAX
 
 
 def sample_from_density(
-    nb_samples, density, method="random", *, dim_compensation="auto"
-):
+    nb_samples: int,
+    density: NDArray,
+    method: Literal["random", "lloyd"] = "random",
+    *,
+    dim_compensation: Literal["auto"] | bool = "auto",
+) -> NDArray:
     """
     Sample points based on a given density distribution.
 
@@ -19,14 +29,14 @@ def sample_from_density(
     ----------
     nb_samples : int
         The number of samples to draw.
-    density : np.ndarray
+    density : NDArray
         An array representing the density distribution from which samples are drawn,
         normalized automatically by its sum during the call for convenience.
-    method : str, optional
+    method : Literal["random", "lloyd"], optional
         The sampling method to use, either 'random' for random sampling over
         the discrete grid defined by the density or 'lloyd' for Lloyd's
         algorithm over a continuous space, by default "random".
-    dim_compensation : str, bool, optional
+    dim_compensation : Literal["auto"], bool, optional
         Whether to apply a specific dimensionality compensation introduced
         in [Cha+14]_. An exponent ``N/(N-1)`` with ``N`` the number of
         dimensions in ``density`` is applied to fix the observed
@@ -37,7 +47,7 @@ def sample_from_density(
 
     Returns
     -------
-    np.ndarray
+    NDArray
         An array of range-normalized sampled locations.
 
     Raises
@@ -50,7 +60,7 @@ def sample_from_density(
     ----------
     .. [Cha+14] Chauffert, Nicolas, Philippe Ciuciu,
        Jonas Kahn, and Pierre Weiss.
-       "Variable density sampling with continuous trajectories."
+       "Variable density sampling with continuous trajectories"
        SIAM Journal on Imaging Sciences 7, no. 4 (2014): 1962-1992.
     """
     try:
@@ -63,7 +73,7 @@ def sample_from_density(
         ) from err
 
     # Define dimension variables
-    shape = np.array(density.shape)
+    shape = density.shape
     nb_dims = len(shape)
     max_nb_samples = np.prod(shape)
     density = density / np.sum(density)
@@ -81,7 +91,7 @@ def sample_from_density(
         density = density / np.sum(density)
 
     # Sample using specified method
-    rng = nr.default_rng()
+    rng = np.random.default_rng()
     if method == "random":
         choices = rng.choice(
             np.arange(max_nb_samples),
@@ -91,7 +101,7 @@ def sample_from_density(
         )
         locations = np.indices(shape).reshape((nb_dims, -1))[:, choices]
         locations = locations.T + 0.5
-        locations = locations / shape[None, :]
+        locations = locations / np.array(shape)[None, :]
         locations = 2 * KMAX * locations - KMAX
     elif method == "lloyd":
         kmeans = (
@@ -100,17 +110,22 @@ def sample_from_density(
             else BisectingKMeans(n_clusters=nb_samples)
         )
         kmeans.fit(
-            np.indices(density.shape).reshape((nb_dims, -1)).T,
+            np.indices(shape).reshape((nb_dims, -1)).T,
             sample_weight=density.flatten(),
         )
[…]
