+++ src/mrinufft/trajectories/inits/random_walk.py
@@ -1,17 +1,19 @@
 """Trajectories based on random walks."""
 
+from typing import Any, Literal
+
 import numpy as np
+from numpy.typing import NDArray
 
 from ..sampling import sample_from_density
 from ..utils import KMAX
 
 
-def _get_adjacent_neighbors_offsets(shape):
-    return np.concatenate([np.eye(len(shape)), -np.eye(len(shape))], axis=0).astype(int)
+def _get_adjacent_neighbors_offsets(nb_dims: int) -> NDArray:
+    return np.concatenate([np.eye(nb_dims), -np.eye(nb_dims)], axis=0).astype(int)
 
 
-def _get_neighbors_offsets(shape):
-    nb_dims = len(shape)
+def _get_neighbors_offsets(nb_dims: int) -> NDArray:
     neighbors = (np.indices([3] * nb_dims) - 1).reshape((nb_dims, -1)).T
     nb_half = neighbors.shape[0] // 2
     # Remove full zero entry
@@ -20,18 +22,25 @@ def _get_neighbors_offsets(shape):
 
 
 def _initialize_ND_random_walk(
-    Nc, Ns, density, *, diagonals=True, pseudo_random=True, **sampling_kwargs
-):
+    Nc: int,
+    Ns: int,
+    density: NDArray,
+    *,
+    diagonals: bool = True,
+    pseudo_random: bool = True,
+    **sampling_kwargs: Any,  # noqa ANN401
+) -> NDArray:
     density = density / np.sum(density)
     flat_density = np.copy(density.flatten())
-    shape = np.array(density.shape)
+    shape = density.shape
+    nb_dims = len(shape)
     mask = np.ones_like(flat_density)
 
     # Prepare neighbor offsets once
     offsets = (
-        _get_neighbors_offsets(shape)
+        _get_neighbors_offsets(nb_dims)
         if diagonals
-        else _get_adjacent_neighbors_offsets(shape)
+        else _get_adjacent_neighbors_offsets(nb_dims)
     )
 
     # Make all random draws at once for performance
@@ -39,9 +48,8 @@ def _initialize_ND_random_walk(
 
     # Initialize shot starting points
     locations = sample_from_density(Nc, density, **sampling_kwargs)
-    choices = np.around((locations + KMAX) * (np.array(density.shape) - 1)).astype(int)
-    choices = np.ravel_multi_index(choices.T, density.shape)
-    # choices = np.random.choice(np.arange(len(flat_density)), size=Nc, p=flat_density)
+    choices = np.around((locations + KMAX) * (np.array(shape) - 1)).astype(int)
+    choices = np.ravel_multi_index(choices.T, shape)
     routes = [choices]
 
     # Walk
@@ -52,7 +60,7 @@ def _initialize_ND_random_walk(
 
         # Find out-of-bound neighbors and ignore them
         invalids = (neighbors < 0).any(axis=0) | (
-            neighbors >= shape[:, None, None]
+            neighbors >= np.array(shape)[:, None, None]
         ).any(axis=0)
         neighbors[:, invalids] = 0
         invalids = invalids.T
@@ -71,25 +79,30 @@ def _initialize_ND_random_walk(
         choices = neighbors[np.arange(Nc), indices]
         routes.append(choices)
 
-        # Update density to account for already drawed positions
+        # Update density to account for already drawn positions
         if pseudo_random:
             flat_density[choices] = (
                 mask[choices] * flat_density[choices] / (mask[choices] + 1)
             )
             mask[choices] += 1
-    routes = np.array(routes).T
 
     # Create trajectory from routes
     locations = np.indices(shape)
-    locations = locations.reshape((len(shape), -1))
[…]
