+++ src/mrinufft/trajectories/maths/tsp_solver.py
@@ -1,29 +1,31 @@
 """Solver for the Travelling Salesman Problem."""
 
 import numpy as np
+from numpy.typing import NDArray
 
 
-def solve_tsp_with_2opt(locations, improvement_threshold=1e-8):
+def solve_tsp_with_2opt(
+    locations: NDArray, improvement_threshold: float = 1e-8
+) -> NDArray:
     """Solve the TSP problem using a 2-opt approach.
 
     A sub-optimal solution to the Travelling Salesman Problem (TSP)
     is provided using the 2-opt approach in O(n²) where chunks of
     an initially random route are reversed, and selected if the
-    total distance is reduced. As a result the route solution
+    total distance is reduced. As a result, the route solution
     does not cross its own path in 2D.
 
     Parameters
     ----------
-    locations : array_like
-        An array of N points with shape (N, D) with D
-        the space dimension.
-    improvement_threshold: float
-                Threshold used as progress criterion to stop the optimization
-                process.
+    locations : NDArray
+        An array of N points with shape (N, D) where D is the space dimension.
+    improvement_threshold : float, optional
+        Threshold used as progress criterion to stop the optimization process.
+        The default is 1e-8.
 
     Returns
     -------
-    array_like
+    NDArray
         The new positions order of shape (N,).
     """
     route = np.arange(locations.shape[0])
