+++ _sources/generated/autoexamples/GPU/example_learn_samples_multicoil.rst
@@ -334,9 +334,9 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466064]/volatile/github-ci-mind-inria/gpu_runner2/_work/_tool/Python/3.10.16/x64/lib/python3.10/site-packages/mrinufft/operators/autodiff.py:98: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
+      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.1545513]/volatile/github-ci-mind-inria/gpu_runner2/_work/_tool/Python/3.10.16/x64/lib/python3.10/site-packages/mrinufft/operators/autodiff.py:98: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
       grad_traj = torch.transpose(torch.sum(grad_traj, dim=1), 0, 1).to(
-      1%|          | 1/100 [00:00<01:15,  1.31steps/s, loss=0.15466064]      1%|          | 1/100 [00:00<01:15,  1.31steps/s, loss=0.27207538]      2%|▏         | 2/100 [00:01<01:14,  1.32steps/s, loss=0.27207538]      2%|▏         | 2/100 [00:01<01:14,  1.32steps/s, loss=0.21596128]      3%|▎         | 3/100 [00:02<01:13,  1.32steps/s, loss=0.21596128]      3%|▎         | 3/100 [00:02<01:13,  1.32steps/s, loss=0.12376805]      4%|▍         | 4/100 [00:03<01:14,  1.29steps/s, loss=0.12376805]      4%|▍         | 4/100 [00:03<01:14,  1.29steps/s, loss=0.10364467]      5%|▌         | 5/100 [00:04<01:21,  1.16steps/s, loss=0.10364467]      5%|▌         | 5/100 [00:04<01:21,  1.16steps/s, loss=0.09531711]      6%|▌         | 6/100 [00:04<01:19,  1.18steps/s, loss=0.09531711]      6%|▌         | 6/100 [00:05<01:19,  1.18steps/s, loss=0.07875554]      7%|▋         | 7/100 [00:05<01:20,  1.16steps/s, loss=0.07875554]      7%|▋         | 7/100 [00:06<01:20,  1.16steps/s, loss=0.083123446]      8%|▊         | 8/100 [00:06<01:20,  1.15steps/s, loss=0.083123446]      8%|▊         | 8/100 [00:06<01:20,  1.15steps/s, loss=0.07696223]       9%|▉         | 9/100 [00:07<01:17,  1.17steps/s, loss=0.07696223]      9%|▉         | 9/100 [00:07<01:17,  1.17steps/s, loss=0.06757349]     10%|█         | 10/100 [00:08<01:16,  1.18steps/s, loss=0.06757349]     10%|█         | 10/100 [00:08<01:16,  1.18steps/s, loss=0.064348236]     11%|█         | 11/100 [00:09<01:16,  1.17steps/s, loss=0.064348236]     11%|█         | 11/100 [00:09<01:16,  1.17steps/s, loss=0.059588775]     12%|█▏        | 12/100 [00:10<01:13,  1.19steps/s, loss=0.059588775]     12%|█▏        | 12/100 [00:10<01:13,  1
[…]
