### 2025-01-13 4602e72
Fix PEP and add details of the quaternion
