### 2025-01-13 0cdfef4
feat:  get right Orientation from raw data (#230)
--------

Co-authored-by: chaithyagr <chaithyagr@gitlab.com>
Co-authored-by: Pierre-antoine Comby <pierre-antoine.comby@crans.org>
