+++ _sources/generated/autoexamples/GPU/example_fastMRI_UNet.rst
@@ -456,9 +456,9 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=1.12]/volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/GPU/example_fastMRI_UNet.py:104: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
+      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=1]/volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/GPU/example_fastMRI_UNet.py:104: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
       axs[0].imshow(np.abs(mri_2D[0]), cmap="gray")
-      1%|          | 1/100 [00:00<01:38,  1.00steps/s, loss=1.12]      1%|          | 1/100 [00:01<01:38,  1.00steps/s, loss=1.01]      2%|▏         | 2/100 [00:01<01:31,  1.07steps/s, loss=1.01]      2%|▏         | 2/100 [00:02<01:31,  1.07steps/s, loss=0.896]      3%|▎         | 3/100 [00:02<01:28,  1.10steps/s, loss=0.896]      3%|▎         | 3/100 [00:02<01:28,  1.10steps/s, loss=0.807]      4%|▍         | 4/100 [00:03<01:26,  1.11steps/s, loss=0.807]      4%|▍         | 4/100 [00:03<01:26,  1.11steps/s, loss=0.742]      5%|▌         | 5/100 [00:04<01:25,  1.11steps/s, loss=0.742]      5%|▌         | 5/100 [00:04<01:25,  1.11steps/s, loss=0.694]      6%|▌         | 6/100 [00:05<01:23,  1.12steps/s, loss=0.694]      6%|▌         | 6/100 [00:05<01:23,  1.12steps/s, loss=0.653]      7%|▋         | 7/100 [00:06<01:31,  1.02steps/s, loss=0.653]      7%|▋         | 7/100 [00:06<01:31,  1.02steps/s, loss=0.613]      8%|▊         | 8/100 [00:07<01:27,  1.05steps/s, loss=0.613]      8%|▊         | 8/100 [00:07<01:27,  1.05steps/s, loss=0.578]      9%|▉         | 9/100 [00:08<01:25,  1.07steps/s, loss=0.578]      9%|▉         | 9/100 [00:08<01:25,  1.07steps/s, loss=0.547]     10%|█         | 10/100 [00:09<01:21,  1.10steps/s, loss=0.547]     10%|█         | 10/100 [00:09<01:21,  1.10steps/s, loss=0.519]     11%|█         | 11/100 [00:10<01:20,  1.10steps/s, loss=0.519]     11%|█         | 11/100 [00:10<01:20,  1.10steps/s, loss=0.494]     12%|█▏        | 12/100 [00:11<01:20,  1.09steps/s, loss=0.494]     12%|█▏        | 12/100 [00:11<01:20,  1.09steps/s, loss=0.471]     13%|█▎        | 13/100 [00:11<01:19,  1.10steps/s, loss=0.471]     13%|█▎        | 13/100 [00:12
[…]

[…]
