+++ _sources/generated/autoexamples/GPU/example_density.rst
@@ -103,7 +103,7 @@ Create sample data
     /volatile/github-ci-mind-inria/gpu_runner2/_work/_tool/Python/3.10.16/x64/lib/python3.10/site-packages/finufft/_interfaces.py:329: UserWarning: Argument `data` does not satisfy the following requirement: C. Copying array (this may reduce performance)
       warnings.warn(f"Argument `{name}` does not satisfy the following requirement: {prop}. Copying array (this may reduce performance)")
 
-    <matplotlib.image.AxesImage object at 0x764d36ce7460>
+    <matplotlib.image.AxesImage object at 0x78ed4a141780>
 
 
 
@@ -331,7 +331,7 @@ Pipe's method is an iterative scheme, that use the interpolation and spreading k
 
  .. code-block:: none
 
-    [0.00154912 0.00706981 0.01425769 ... 0.56895894 0.46830535 0.60650784]
+    [0.00885258 0.04039435 0.08038493 ... 3.2315485  2.6598606  3.4448173 ]
 
 
 
@@ -339,7 +339,7 @@ Pipe's method is an iterative scheme, that use the interpolation and spreading k
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 4.683 seconds)
+   **Total running time of the script:** (0 minutes 4.266 seconds)
 
 
 .. _sphx_glr_download_generated_autoexamples_GPU_example_density.py:

[…]
