### 2025-01-23 82dfd12
feat: add 1d case and stacking.
