+++ examples/GPU/example_3d_trajectory_display.py
@@ -39,7 +39,12 @@ def plot_slices(axs, volume, title=""):
         set_labels(axs[i], i)
 
 
-def create_grid(grid_type, trajectories, **kwargs):
+
+# %%
+# Helper function to showcase the features of `get_gridded_trajectory` function
+# This function will first grid the trajectory using the `get_gridded_trajectory` 
+# function and then plot the mid-plane slices of the gridded trajectory.
+def create_grid(grid_type, trajectories, traj_params, **kwargs):
     fig, axs = plt.subplots(3, 3, figsize=(10, 10))
     for i, (name, traj) in enumerate(trajectories.items()):
         grid = get_gridded_trajectory(
@@ -68,7 +73,7 @@ traj_params = {
 
 # %%
 # Display the density of the trajectories, along the 3 mid-planes. For this, make `grid_type="density"`.
-create_grid("density", trajectories)
+create_grid("density", trajectories, traj_params)
 plt.suptitle("Sampling Density")
 plt.show()
 
@@ -76,36 +81,39 @@ plt.show()
 # %%
 # Display the sampling times over the trajectories. For this, make `grid_type="time"`.
 # It helps to check the sampling times over the k-space trajectories, which can be responsible for excessive off-resonance artifacts.
-create_grid("time", trajectories)
+# Note that this is just a relative visualization of sample times on a colour scale, and the actual sampling time.
+create_grid("time", trajectories, traj_params)
 plt.suptitle("Sampling Time")
 plt.show()
 
 # %%
 # Display the inversion time of the trajectories. For this, make `grid_type="inversion"`.
-# This helps in obtaining the inversion time when particular region of k-space is sampled, assuming the trajectories are time ordered.
-# This helps understand any issues for imaging involving inversion recovery.
-# The argument `turbo_factor` can be used to tell what is the number of echoes between 2 inversion pulses.
-create_grid("inversion", trajectories, turbo_factor=64)
+# This helps in obtaining the inversion time when particular region of k-space is sampled, assuming the trajectories are time ordered, 
+# and the argument `turbo_factor` is specified, which is time between 2 inversion pulses. 
+create_grid("inversion", trajectories, traj_params, turbo_factor=64)
 plt.suptitle("Inversion Time")
 plt.show()
 # %%
 # Display the k-space holes in the trajectories. For this, make `grid_type="holes"`.
-# This helps in understanding the k-space holes, and can help debug artifacts in reconstructions.
-create_grid("holes", trajectories, threshold=1e-2)
+# K-space holes can occur in a sampling trajectory when it is optimized or learned based on a specific loss.
+# This feature can be used to identify the k-space holes, which could lead to Gibbs like ringing artifacts in the images.
+create_grid("holes", trajectories, traj_params, threshold=1e-2)
 plt.suptitle("K-space Holes")
 plt.show()
 # %%
 # Display the gradient strength of the trajectories. For this, make `grid_type="gradients"`.
-# This helps in understanding the gradient strength applied at specific k-space region.
-# This can also be used as a surrogate to k-space "velocity", i.e. how fast does trajectory pass through a given region in k-space
-create_grid("gradients", trajectories)
+# This helps in displaying the gradient strength applied at specific k-space region, 
+# which can be used as a surrogate to k-space "velocity", i.e. how fast does trajectory pass through a given region in k-space.
+# It could be useful while characterizing spatial SNR profile in k-space
+create_grid("gradients", trajectories, traj_params)
 plt.suptitle("Gradient Strength")
 plt.show()
 
 # %%
 # Display the slew rates of the trajectories. For this, make `grid_type="slew"`.
-# This helps in understanding the slew rates applied at specific k-space region.
[…]
