### 2025-01-30 dc0ccf8
feat: add typing and remove typos
