### 2025-02-02 #7501 @Kevin-Mattheus-Moerman
@alexjaffray, @spinicist, @cncastillo, @mathieuboudreau, @felixhorger, @emilljungberg, @imr-framework, @bwheelz36, @curtcorum, would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is a Python library titled `MRI-NUFFT: Doing non-Cartesian MRI has never been easier`

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/mind-inria/mri-nufft) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07501/joss.07501/10.21105.joss.07501.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of an ongoing review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!

Kevin Moerman
_Associated Editor in Chief, Journal of Open Source Software_
