+++ _sources/generated/autoexamples/GPU/example_learn_samples_multicoil.rst
@@ -334,9 +334,9 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466063]/volatile/github-ci-mind-inria/gpu_runner2/_work/_tool/Python/3.10.16/x64/lib/python3.10/site-packages/mrinufft/operators/autodiff.py:98: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
+      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466073]/volatile/github-ci-mind-inria/gpu_runner2/_work/_tool/Python/3.10.16/x64/lib/python3.10/site-packages/mrinufft/operators/autodiff.py:98: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
       grad_traj = torch.transpose(torch.sum(grad_traj, dim=1), 0, 1).to(
-      1%|          | 1/100 [00:00<01:19,  1.25steps/s, loss=0.15466063]      1%|          | 1/100 [00:00<01:19,  1.25steps/s, loss=0.27207497]      2%|▏         | 2/100 [00:01<01:18,  1.26steps/s, loss=0.27207497]      2%|▏         | 2/100 [00:01<01:18,  1.26steps/s, loss=0.2069565]       3%|▎         | 3/100 [00:02<01:14,  1.30steps/s, loss=0.2069565]      3%|▎         | 3/100 [00:02<01:14,  1.30steps/s, loss=0.10875603]      4%|▍         | 4/100 [00:03<01:13,  1.30steps/s, loss=0.10875603]      4%|▍         | 4/100 [00:03<01:13,  1.30steps/s, loss=0.10856793]      5%|▌         | 5/100 [00:04<01:19,  1.19steps/s, loss=0.10856793]      5%|▌         | 5/100 [00:04<01:19,  1.19steps/s, loss=0.092213795]      6%|▌         | 6/100 [00:04<01:15,  1.24steps/s, loss=0.092213795]      6%|▌         | 6/100 [00:04<01:15,  1.24steps/s, loss=0.078813046]      7%|▋         | 7/100 [00:05<01:13,  1.26steps/s, loss=0.078813046]      7%|▋         | 7/100 [00:05<01:13,  1.26steps/s, loss=0.07408297]       8%|▊         | 8/100 [00:06<01:12,  1.27steps/s, loss=0.07408297]      8%|▊         | 8/100 [00:06<01:12,  1.27steps/s, loss=0.073726356]      9%|▉         | 9/100 [00:07<01:10,  1.28steps/s, loss=0.073726356]      9%|▉         | 9/100 [00:07<01:10,  1.28steps/s, loss=0.06729086]      10%|█         | 10/100 [00:07<01:08,  1.31steps/s, loss=0.06729086]     10%|█         | 10/100 [00:08<01:08,  1.31steps/s, loss=0.066663474]     11%|█         | 11/100 [00:08<01:07,  1.32steps/s, loss=0.066663474]     11%|█         | 11/100 [00:08<01:07,  1.32steps/s, loss=0.062695906]     12%|█▏        | 12/100 [00:09<01:05,  1.34steps/s, loss=0.062695906]     12%|█▏        | 12/100 [00:09<01:05,
[…]
