### 2025-02-10 56601e6
fix: set the computed density using the setter.
