+++ src/mrinufft/operators/base.py
@@ -10,18 +10,15 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from functools import partial
-
+from typing import ClassVar, Callable
 import numpy as np
+from numpy.typing import NDArray
 
 from mrinufft._array_compat import with_numpy, with_numpy_cupy, AUTOGRAD_AVAILABLE
 from mrinufft._utils import auto_cast, power_method
 from mrinufft.density import get_density
 from mrinufft.extras import get_smaps
 from mrinufft.operators.interfaces.utils import is_cuda_array, is_host_array
-
-if AUTOGRAD_AVAILABLE:
-    from mrinufft.operators.autodiff import MRINufftAutoGrad
-
 
 # Mapping between numpy float and complex types.
 DTYPE_R2C = {"float32": "complex64", "float64": "complex128"}
@@ -122,6 +119,9 @@ class FourierOperatorBase(ABC):
     _grad_wrt_data = False
     _grad_wrt_traj = False
 
+    backend: ClassVar[str]
+    available: ClassVar[bool]
+
     def __init__(self):
         if not self.available:
             raise RuntimeError(f"'{self.backend}' backend is not available.")
@@ -207,21 +207,21 @@ class FourierOperatorBase(ABC):
         """
         pass
 
-    def data_consistency(self, image, obs_data):
+    def data_consistency(self, image_data, obs_data):
         """Compute the gradient data consistency.
 
         This is the naive implementation using adj_op(op(x)-y).
         Specific backend can (and should!) implement a more efficient version.
         """
-        return self.adj_op(self.op(image) - obs_data)
+        return self.adj_op(self.op(image_data) - obs_data)
 
     def with_off_resonance_correction(self, B, C, indices):
         """Return a new operator with Off Resonnance Correction."""
-        from ..off_resonance import MRIFourierCorrected
+        from .off_resonance import MRIFourierCorrected
 
         return MRIFourierCorrected(self, B, C, indices)
 
-    def compute_smaps(self, method=None):
+    def compute_smaps(self, method: NDArray | Callable | str | dict | None = None):
         """Compute the sensitivity maps and set it.
 
         Parameters
@@ -285,6 +285,8 @@ class FourierOperatorBase(ABC):
             raise ValueError("Autograd not available, ensure torch is installed.")
         if not self.autograd_available:
             raise ValueError("Backend does not support auto-differentiation.")
+
+        from mrinufft.operators.autodiff import MRINufftAutoGrad
 
         return MRINufftAutoGrad(self, wrt_data=wrt_data, wrt_traj=wrt_traj)
 
@@ -401,9 +403,9 @@ class FourierOperatorBase(ABC):
         return self._smaps
 
     @smaps.setter
-    def smaps(self, smaps):
-        self._check_smaps_shape(smaps)
-        self._smaps = smaps
+    def smaps(self, new_smaps):
+        self._check_smaps_shape(new_smaps)
+        self._smaps = new_smaps
 
     def _check_smaps_shape(self, smaps):
         """Check the shape of the sensitivity maps."""
@@ -421,13 +423,13 @@ class FourierOperatorBase(ABC):
         return self._density
 
     @density.setter
-    def density(self, density):
-        if density is None:
+    def density(self, new_density):
+        if new_density is None:
             self._density = None
-        elif len(density) != self.n_samples:
+        elif len(new_density) != self.n_samples:
             raise ValueError("Density and samples should have the same length")
         else:
-            self._density = density
+            self._density = new_density
 
     @property
     def dtype(self):
@@ -435,8 +437,8 @@ class FourierOperatorBase(ABC):
         return self._dtype
 
     @dtype.setter
[…]
