+++ src/mrinufft/operators/interfaces/cufinufft.py
@@ -26,7 +26,6 @@ try:
 except ImportError:
     CUFINUFFT_AVAILABLE = False
 
-
 OPTS_FIELD_DECODE = {
     "gpu_method": {1: "nonuniform pts driven", 2: "shared memory"},
     "gpu_sort": {0: "no sort (GM)", 1: "sort (GM-sort)"},
@@ -269,10 +268,12 @@ class MRICufiNUFFT(FourierOperatorBase):
             self._smaps = new_smaps
 
     @FourierOperatorBase.samples.setter
-    def samples(self, samples):
+    def samples(self, new_samples):
         """Update the plans when changing the samples."""
         self._samples = np.asfortranarray(
-            proper_trajectory(samples, normalize="pi").astype(np.float32, copy=False)
+            proper_trajectory(new_samples, normalize="pi").astype(
+                np.float32, copy=False
+            )
         )
         for typ in [1, 2, "grad"]:
             if typ == "grad" and not self._grad_wrt_traj:
