+++ _sources/generated/autoexamples/GPU/example_learn_samples.rst
@@ -282,7 +282,7 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.29853112]      1%|          | 1/100 [00:00<01:00,  1.64steps/s, loss=0.29853112]      1%|          | 1/100 [00:00<01:00,  1.64steps/s, loss=0.37153855]      2%|▏         | 2/100 [00:01<01:01,  1.59steps/s, loss=0.37153855]      2%|▏         | 2/100 [00:01<01:01,  1.59steps/s, loss=0.32791892]      3%|▎         | 3/100 [00:01<01:02,  1.56steps/s, loss=0.32791892]      3%|▎         | 3/100 [00:01<01:02,  1.56steps/s, loss=0.2729209]       4%|▍         | 4/100 [00:02<01:01,  1.55steps/s, loss=0.2729209]      4%|▍         | 4/100 [00:02<01:01,  1.55steps/s, loss=0.285618]       5%|▌         | 5/100 [00:03<01:05,  1.46steps/s, loss=0.285618]      5%|▌         | 5/100 [00:03<01:05,  1.46steps/s, loss=0.25701326]      6%|▌         | 6/100 [00:03<01:01,  1.54steps/s, loss=0.25701326]      6%|▌         | 6/100 [00:03<01:01,  1.54steps/s, loss=0.25487423]      7%|▋         | 7/100 [00:04<00:58,  1.59steps/s, loss=0.25487423]      7%|▋         | 7/100 [00:04<00:58,  1.59steps/s, loss=0.24471316]      8%|▊         | 8/100 [00:05<00:56,  1.62steps/s, loss=0.24471316]      8%|▊         | 8/100 [00:05<00:56,  1.62steps/s, loss=0.23612848]      9%|▉         | 9/100 [00:05<00:55,  1.65steps/s, loss=0.23612848]      9%|▉         | 9/100 [00:05<00:55,  1.65steps/s, loss=0.23718472]     10%|█         | 10/100 [00:06<00:53,  1.67steps/s, loss=0.23718472]     10%|█         | 10/100 [00:06<00:53,  1.67steps/s, loss=0.22051743]     11%|█         | 11/100 [00:06<00:53,  1.68steps/s, loss=0.22051743]     11%|█         | 11/100 [00:06<00:53,  1.68steps/s, loss=0.22469674]     12%|█▏        | 12/100 [00:07<00:53,  1.65steps/s, loss=0.22469674]     12%|█▏        | 12/100 [00:07<00:53,  1.65steps/s, loss=0.21266262]     13%|█▎        | 13/100 [00:08<00:52,  1.65steps/s, loss=0.21266262]     13%|█▎        | 13/100 [00:08<00:52,  1.65steps/s, loss=0.21103373]     14%|█▍        | 14/100 [00:08<00:57,  1.48steps/s, loss=0.21103373]     14%|█▍        | 14/100 [00:08<00:57,  1.48steps/s, loss=0.20615734]     15%|█▌        | 15/100 [00:09<00:55,  1.53steps/s, loss=0.20615734]     15%|█▌        | 15/100 [00:09<00:55,  1.53steps/s, loss=0.20250212]     16%|█▌        | 16/100 [00:10<00:53,  1.56steps/s, loss=0.20250212]     16%|█▌        | 16/100 [00:10<00:53,  1.56steps/s, loss=0.2019312]      17%|█▋        | 17/100 [00:10<00:52,  1.58steps/s, loss=0.2019312]     17%|█▋      
[…]
