+++ _sources/generated/_autosummary/mrinufft.operators.off_resonance.MRIFourierCorrected.rst
@@ -51,5 +51,7 @@ MRIFourierCorrected
       ~MRIFourierCorrected.smaps
       ~MRIFourierCorrected.uses_density
       ~MRIFourierCorrected.uses_sense
+      ~MRIFourierCorrected.backend
+      ~MRIFourierCorrected.available
    
    
