+++ _sources/generated/_autosummary/mrinufft.operators.interfaces.torchkbnufft.MRITorchKbNufft.rst
@@ -52,5 +52,6 @@ MRITorchKbNufft
       ~MRITorchKbNufft.smaps
       ~MRITorchKbNufft.uses_density
       ~MRITorchKbNufft.uses_sense
+      ~MRITorchKbNufft.backend
    
    
