+++ src/mrinufft/operators/interfaces/finufft.py
@@ -200,12 +200,10 @@ class MRIfinufft(FourierOperatorCPU):
             raise ValueError(
                 "finufft is not available, cannot estimate the density compensation"
             )
-        original_shape = volume_shape
-        volume_shape = np.array([_next235beven(int(osf * i), 1) for i in volume_shape])
         grid_op = MRIfinufft(
             samples=kspace_loc,
             shape=volume_shape,
-            upsampfac=1,
+            upsampfac=osf,
             spreadinterponly=1,
             spread_kerevalmeth=0,
             **kwargs,
@@ -218,8 +216,8 @@ class MRIfinufft(FourierOperatorCPU):
                 ).squeeze()
             )
         if normalize:
-           test_op = MRIfinufft(samples=kspace_loc, shape=original_shape, **kwargs)
-           test_im = np.ones(original_shape, dtype=test_op.cpx_dtype)
+           test_op = MRIfinufft(samples=kspace_loc, shape=volume_shape, **kwargs)
+           test_im = np.ones(volume_shape, dtype=test_op.cpx_dtype)
            test_im_recon = test_op.adj_op(density_comp * test_op.op(test_im))
            density_comp /= np.mean(np.abs(test_im_recon))
         return density_comp.squeeze()
