+++ src/mrinufft/operators/interfaces/cufinufft.py
@@ -881,12 +881,10 @@ class MRICufiNUFFT(FourierOperatorBase):
             raise ValueError(
                 "cufinufft is not available, cannot estimate the density compensation"
             )
-        original_shape = volume_shape
-        volume_shape = np.array([_next235beven(int(osf * i), 1) for i in volume_shape])
         grid_op = MRICufiNUFFT(
             samples=kspace_loc,
             shape=volume_shape,
-            upsampfac=1,
+            upsampfac=osf,
             gpu_spreadinterponly=1,
             gpu_kerevalmeth=0,
             **kwargs,
@@ -899,8 +897,8 @@ class MRICufiNUFFT(FourierOperatorBase):
                 ).squeeze()
             )
         if normalize:
-            test_op = MRICufiNUFFT(samples=kspace_loc, shape=original_shape, **kwargs)
-            test_im = cp.ones(original_shape, dtype=test_op.cpx_dtype)
+            test_op = MRICufiNUFFT(samples=kspace_loc, shape=volume_shape, **kwargs)
+            test_im = cp.ones(volume_shape, dtype=test_op.cpx_dtype)
             test_im_recon = test_op.adj_op(density_comp * test_op.op(test_im))
             density_comp /= cp.mean(cp.abs(test_im_recon))
         return density_comp.squeeze()
