+++ tests/operators/test_density_for_op.py
@@ -17,7 +17,7 @@ def radial_distance(traj, shape):
     return weights
 
 
-@parametrize("osf", [1, 1.25, 2])
+@parametrize("osf", [1, 1.5, 2])
 @parametrize_with_cases(
     "traj, shape",
     cases=[
@@ -31,6 +31,8 @@ def test_pipe(backend, traj, shape, osf):
     distance = radial_distance(traj, shape)
     if osf != 2 and backend == "tensorflow":
         pytest.skip("OSF < 2 not supported for tensorflow.")
+    if osf == 1 and 'finufft' in backend:
+        pytest.skip("cufinufft and finufft dont support OSF=1")
     result = pipe(traj, shape, backend=backend, osf=osf, num_iterations=10)
     if backend == "cufinufft":
         result = result.get()
@@ -41,12 +43,9 @@ def test_pipe(backend, traj, shape, osf):
     if osf == 2:
         r_err = 0.1
         slope_err = 0.1
-    if "finufft" in backend:
-        r_err *= 2
-        slope_err = slope_err * 2 if slope_err is not None else None
-        if len(shape) == 3:
-            r_err *= 2
-            slope_err = slope_err * 2 if slope_err is not None else None
+    if 'finufft' in backend:
+        r_err *= 3
+        slope_err = slope_err * 4 if slope_err is not None else None
     elif backend == "tensorflow":
         r_err = 0.5
     assert_correlate(result, distance, slope=1, slope_err=slope_err, r_value_err=r_err)
