### 2025-02-11 eb11ef1
Update src/mrinufft/trajectories/display3D.py
Co-authored-by: Guillaume Daval-Frérot <gdavalfrerot@gmail.com>
