### 2025-02-11 7aea83f
Update src/mrinufft/trajectories/display3D.py
Co-authored-by: Guillaume Daval-Frérot <gdavalfrerot@gmail.com>
