### 2025-02-11 842c793
Update examples/GPU/example_3d_trajectory_display.py
Co-authored-by: Guillaume Daval-Frérot <gdavalfrerot@gmail.com>
