### 2025-02-12 5aa3fdb
Update src/mrinufft/trajectories/display3D.py
Co-authored-by: Guillaume Daval-Frérot <gdavalfrerot@gmail.com>
