### 2025-02-12 c0004bc
Update examples/GPU/example_3d_trajectory_display.py
Co-authored-by: Guillaume Daval-Frérot <gdavalfrerot@gmail.com>
