+++ examples/GPU/example_3d_trajectory_display.py
@@ -59,7 +59,7 @@ def create_grid(grid_type, trajectories, traj_params, **kwargs):
 
 
 # %%
-# Create a bunch of sample trajectories
+# Create a bunch of sampling trajectories
 trajectories = {
     "Radial": mtt.initialize_3D_phyllotaxis_radial(64 * 8, 64),
     "FLORET": mtt.initialize_3D_floret(64 * 8, 64, nb_revolutions=2),
