### 2025-02-12 7ddd828
[docs] final updates and fixes and also try to fix #235
