+++ examples/GPU/example_3d_trajectory_display.py
@@ -1,7 +1,7 @@
 """
-======================
-3D Trajectory display
-======================
+=============================
+3D Trajectory Gridded display
+=============================
 
 In this example, we show some tools available to display 3D trajectories.
 It can be used to understand the k-space sampling patterns, visualize the trajectories, see the sampling times, gradient strengths, slew rates etc.
@@ -19,7 +19,8 @@ import numpy as np
 
 
 # %%
-
+# Helper function to Displaying 3D Gridded Trajectories
+# =====================================================
 # Utility function to plot mid-plane slices for 3D volumes
 def plot_slices(axs, volume, title=""):
     def set_labels(ax, axis_num=None):
@@ -41,11 +42,14 @@ def plot_slices(axs, volume, title=""):
 
 
 # %%
+# Helper function to Displaying 3D Trajectories
+# =============================================
 # Helper function to showcase the features of `get_gridded_trajectory` function
 # This function will first grid the trajectory using the `get_gridded_trajectory`
 # function and then plot the mid-plane slices of the gridded trajectory.
 def create_grid(grid_type, trajectories, traj_params, **kwargs):
     fig, axs = plt.subplots(3, 3, figsize=(10, 10))
+    plt.subplots_adjust(hspace=0.5)
     for i, (name, traj) in enumerate(trajectories.items()):
         grid = get_gridded_trajectory(
             traj,
@@ -55,10 +59,11 @@ def create_grid(grid_type, trajectories, traj_params, **kwargs):
             **kwargs,
         )
         plot_slices(axs[:, i], grid, title=name)
-    plt.tight_layout()
 
 
 # %%
+# Trajectories to display
+# =======================
 # Create a bunch of sampling trajectories
 trajectories = {
     "Radial": mtt.initialize_3D_phyllotaxis_radial(64 * 8, 64),
@@ -72,48 +77,60 @@ traj_params = {
 }
 
 # %%
+# Sampling density
+# =================
 # Display the density of the trajectories, along the 3 mid-planes. For this, make `grid_type="density"`.
 create_grid("density", trajectories, traj_params)
-plt.suptitle("Sampling Density")
+plt.suptitle("Sampling Density", y=0.98, x=0.52, fontsize=20)
 plt.show()
 
 
 # %%
+# Sampling time
+# =============
 # Display the sampling times over the trajectories. For this, make `grid_type="time"`.
 # It helps to check the sampling times over the k-space trajectories, which can be responsible for excessive off-resonance artifacts.
 # Note that this is just a relative visualization of sample times on a colour scale, and the actual sampling time.
 create_grid("time", trajectories, traj_params)
-plt.suptitle("Sampling Time")
+plt.suptitle("Sampling Time", y=0.98, x=0.52, fontsize=20)
 plt.show()
 
 # %%
+# Inversion time
+# ==============
 # Display the inversion time of the trajectories. For this, make `grid_type="inversion"`.
 # This helps in obtaining the inversion time when particular region of k-space is sampled, assuming the trajectories are time ordered,
 # and the argument `turbo_factor` is specified, which is the time between 2 inversion pulses.
 create_grid("inversion", trajectories, traj_params, turbo_factor=64)
-plt.suptitle("Inversion Time")
+plt.suptitle("Inversion Time", y=0.98, x=0.52, fontsize=20)
 plt.show()
 # %%
+# K-space holes
+# =============
 # Display the k-space holes in the trajectories. For this, make `grid_type="holes"`.
 # K-space holes are areas with missing trajectory coverage, and can typically occur with learning-based trajectories when optimized using a specific loss.
[…]
