### 2025-02-12 d12c242
Update src/mrinufft/trajectories/display3D.py
Co-authored-by: Guillaume Daval-Frérot <gdavalfrerot@gmail.com>
