+++ _sources/generated/autoexamples/GPU/example_learn_samples_multicoil.rst
@@ -334,9 +334,9 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466064]/volatile/github-ci-mind-inria/gpu_runner2/_work/_tool/Python/3.10.16/x64/lib/python3.10/site-packages/mrinufft/operators/autodiff.py:98: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
+      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466063]/volatile/github-ci-mind-inria/gpu_runner2/_work/_tool/Python/3.10.16/x64/lib/python3.10/site-packages/mrinufft/operators/autodiff.py:98: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
       grad_traj = torch.transpose(torch.sum(grad_traj, dim=1), 0, 1).to(
-      1%|          | 1/100 [00:01<01:52,  1.14s/steps, loss=0.15466064]      1%|          | 1/100 [00:01<01:52,  1.14s/steps, loss=0.27207482]      2%|▏         | 2/100 [00:02<02:03,  1.26s/steps, loss=0.27207482]      2%|▏         | 2/100 [00:02<02:03,  1.26s/steps, loss=0.20695417]      3%|▎         | 3/100 [00:03<02:00,  1.24s/steps, loss=0.20695417]      3%|▎         | 3/100 [00:03<02:00,  1.24s/steps, loss=0.10849723]      4%|▍         | 4/100 [00:04<01:59,  1.25s/steps, loss=0.10849723]      4%|▍         | 4/100 [00:05<01:59,  1.25s/steps, loss=0.10862867]      5%|▌         | 5/100 [00:06<02:11,  1.39s/steps, loss=0.10862867]      5%|▌         | 5/100 [00:06<02:11,  1.39s/steps, loss=0.09122055]      6%|▌         | 6/100 [00:07<02:05,  1.33s/steps, loss=0.09122055]      6%|▌         | 6/100 [00:08<02:05,  1.33s/steps, loss=0.078682885]      7%|▋         | 7/100 [00:09<02:00,  1.30s/steps, loss=0.078682885]      7%|▋         | 7/100 [00:09<02:00,  1.30s/steps, loss=0.07434255]       8%|▊         | 8/100 [00:10<01:54,  1.24s/steps, loss=0.07434255]      8%|▊         | 8/100 [00:10<01:54,  1.24s/steps, loss=0.07454668]      9%|▉         | 9/100 [00:11<01:49,  1.20s/steps, loss=0.07454668]      9%|▉         | 9/100 [00:11<01:49,  1.20s/steps, loss=0.0696176]      10%|█         | 10/100 [00:12<01:45,  1.17s/steps, loss=0.0696176]     10%|█         | 10/100 [00:12<01:45,  1.17s/steps, loss=0.06671026]     11%|█         | 11/100 [00:13<01:48,  1.22s/steps, loss=0.06671026]     11%|█         | 11/100 [00:14<01:48,  1.22s/steps, loss=0.061355047]     12%|█▏        | 12/100 [00:15<01:51,  1.26s/steps, loss=0.061355047]     12%|█▏        | 12/100 [00:15<01:51,  1.26s
[…]
