+++ _sources/generated/_autosummary/mrinufft.io.utils.remove_extra_kspace_samples.rst
@@ -1,0 +1,11 @@
+remove\_extra\_kspace\_samples
+==============================
+
+.. currentmodule:: mrinufft.io.utils
+
+.. autofunction:: remove_extra_kspace_samples
+
+..  _sphx_glr_backref_mrinufft.io.utils.remove_extra_kspace_samples:
+
+.. minigallery:: mrinufft.io.utils.remove_extra_kspace_samples
+   :add-heading:
