### 2025-02-14 d741b1e
Add arc length parameterization
