### 2025-02-16 817c380
fixup! Add arc length parameterization
