### 2025-02-19 6b5667c
[docs] this wont work @guillaume, you can merge this. #235 will be handeled separately
