### 2025-02-23 bcc5bdc
Apply suggestions from code review
[docs] final updates

Co-authored-by: Guillaume Daval-Frérot <gdavalfrerot@gmail.com>
