+++ _sources/generated/autoexamples/GPU/example_learn_straight_line_readouts.rst
@@ -332,7 +332,7 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.32962516]      1%|          | 1/100 [00:06<10:14,  6.21s/steps, loss=0.32962516]      1%|          | 1/100 [00:06<10:14,  6.21s/steps, loss=0.3228115]       2%|▏         | 2/100 [00:11<09:37,  5.89s/steps, loss=0.3228115]      2%|▏         | 2/100 [00:12<09:37,  5.89s/steps, loss=0.3229206]      3%|▎         | 3/100 [00:18<09:42,  6.00s/steps, loss=0.3229206]      3%|▎         | 3/100 [00:18<09:42,  6.00s/steps, loss=0.32200104]      4%|▍         | 4/100 [00:23<09:25,  5.89s/steps, loss=0.32200104]      4%|▍         | 4/100 [00:23<09:25,  5.89s/steps, loss=0.26681262]      5%|▌         | 5/100 [00:31<10:39,  6.73s/steps, loss=0.26681262]      5%|▌         | 5/100 [00:32<10:39,  6.73s/steps, loss=0.27667597]      6%|▌         | 6/100 [00:38<10:42,  6.83s/steps, loss=0.27667597]      6%|▌         | 6/100 [00:39<10:42,  6.83s/steps, loss=0.28662091]      7%|▋         | 7/100 [00:46<10:45,  6.94s/steps, loss=0.28662091]      7%|▋         | 7/100 [00:46<10:45,  6.94s/steps, loss=0.29520935]      8%|▊         | 8/100 [00:52<10:08,  6.61s/steps, loss=0.29520935]      8%|▊         | 8/100 [00:52<10:08,  6.61s/steps, loss=0.27312142]      9%|▉         | 9/100 [00:57<09:31,  6.28s/steps, loss=0.27312142]      9%|▉         | 9/100 [00:57<09:31,  6.28s/steps, loss=0.25310266]     10%|█         | 10/100 [01:04<09:50,  6.56s/steps, loss=0.25310266]     10%|█         | 10/100 [01:04<09:50,  6.56s/steps, loss=0.24927798]     11%|█         | 11/100 [01:12<10:03,  6.78s/steps, loss=0.24927798]     11%|█         | 11/100 [01:12<10:03,  6.78s/steps, loss=0.24641837]     12%|█▏        | 12/100 [01:19<10:17,  7.02s/steps, loss=0.24641837]     12%|█▏        | 12/100 [01:19<10:17,  7.02s/steps, loss=0.24721527]     13%|█▎        | 13/100 [01:26<10:10,  7.01s/steps, loss=0.24721527]     13%|█▎        | 13/100 [01:26<10:10,  7.01s/steps, loss=0.2443713]      14%|█▍        | 14/100 [01:33<10:10,  7.10s/steps, loss=0.2443713]     14%|█▍        | 14/100 [01:34<10:10,  7.10s/steps, loss=0.24082421]     15%|█▌        | 15/100 [01:41<10:05,  7.12s/steps, loss=0.24082421]     15%|█▌        | 15/100 [01:41<10:05,  7.12s/steps, loss=0.2385892]      16%|█▌        | 16/100 [01:48<10:00,  7.15s/steps, loss=0.2385892]     16%|█▌        | 16/100 [01:48<10:00,  7.15s/steps, loss=0.24147321]     17%|█▋        | 17/100 [01:55<09:44,  7.04s/steps, loss=0.24147321]     17%|█▋       
[…]
