+++ _sources/generated/autoexamples/example_display_config.rst
@@ -313,7 +313,7 @@ are observed.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 7.523 seconds)
+   **Total running time of the script:** (0 minutes 5.796 seconds)
 
 
 .. _sphx_glr_download_generated_autoexamples_example_display_config.py:
