+++ _sources/generated/_autosummary/mrinufft.trajectories.rst
@@ -26,6 +26,7 @@
    :recursive:
 
    display
+   display3D
    gradients
    inits
    maths
