+++ _sources/generated/_autosummary/mrinufft.trajectories.display3D.get_gridded_trajectory.rst
@@ -1,0 +1,11 @@
+get\_gridded\_trajectory
+========================
+
+.. currentmodule:: mrinufft.trajectories.display3D
+
+.. autofunction:: get_gridded_trajectory
+
+..  _sphx_glr_backref_mrinufft.trajectories.display3D.get_gridded_trajectory:
+
+.. minigallery:: mrinufft.trajectories.display3D.get_gridded_trajectory
+   :add-heading:
