### 2025-03-02 mind-inria/mri-nufft#240 @bwheelz36
not very clear how to call different backends
reviewing for https://github.com/openjournals/joss-reviews/issues/7743#issuecomment-2679948969

was working through basic example:
https://mind-inria.github.io/mri-nufft/index.html

wanted to switch from finufft to pynufft. not clear what string to use, tried `pynufft` `pynufft_cpu` before eventually figuring out it's `pynufft-cpu`.
the `backend_name` does not list what values are supported. I can't find where this is documented? it probably should be in the function that users interact with though
