### 2025-03-02 mind-inria/mri-nufft#241 @bwheelz36
community guidelines
Reviewing for https://github.com/openjournals/joss-reviews/issues/7743#issuecomment-2679948969

checklist asks for community guidelines, I think these are missing
