### 2025-03-02 mind-inria/mri-nufft#241 @paquiteau
Hello @bwheelz36, thank you for your review on MRI-NUFFT !

Our community guidelines are described in the `CONTRIBUTING.md` file. We also have a "Developing MRI-NUFFT" page in our documentation [here](https://mind-inria.github.io/mri-nufft/misc/development.html).
