### 2025-03-02 mind-inria/mri-nufft#239 @paquiteau
Hello, thank for your interest, the kind words and your review of MRI-NUFFT !

I linked the related issue as sub-issue (new github functionality I am discovering, let's see how it goes)

>  why is CI failing on main branch?

it's quite new (CI was all green last week :smiling_face_with_tear:)  and is likely because we use a local CI runner (for computation) see #242.

> [[here](https://mind-inria.github.io/mri-nufft/generated/autoexamples/example_2D_trajectories.html)](https://mind-inria.github.io/mri-nufft/generated/autoexamples/example_2D_trajectories.html) I'm pretty sure the internal / external examples are the wrong way around. could be elsew[here](https://mind-inria.github.io/mri-nufft/mrinufft_convention.html) too

Could you please clarify what you mean by internal /external examples?  The last link you provided is a general description of our interface, not an example (and therefore in our Explanation section of the documentation)

> The below lines of code are a bit unconventional, in that I first get use get_operator to ... effectively import a class I think? and then below that I instantiate the class ?

Indeed, but we should update our example to be more straight-forward. The recommended usage would rather be:

```python
nufft = get_operator("finufft", samples, shape, density="voronoi", n_coils=1)

ksp = nufft.op(image) # get the k-space
image_adjoint = nufft.adj_op(ksp) # get the adjoint of the kspace

```
Which directly builds the operator and configures it (here to use "voronoi" density compensation, and a single coil)
