### 2025-03-03 603eb2c
feat: clarify use of operator backend.
fix: #240
