+++ .github/workflows/test-ci.yml
@@ -16,6 +16,9 @@ env:
   PYTHON_VERSION: "3.10"
   BART_VERSION: "0.8.00"
   ref_backend: "finufft"
+  create_venv: "python -m venv $RUNNER_WORKSPACE/venv/"
+  activate_venv_venv: "source $RUNNER_WORKSPACE/venv/bin/activate"
+  setup_cuda: "export CUDA_BIN_PATH=/usr/local/cuda-12.1/;export PATH=/usr/local/cuda-12.1/bin/:${PATH};export LD_LIBRARY_PATH=/usr/local/cuda-12.1/lib64/:${LD_LIBRARY_PATH}"
 
 jobs:
   test-cpu:
@@ -44,6 +47,9 @@ jobs:
       - name: Install Python Deps
         shell: bash
         run: |
+          python --version
+          ${{env.create_venv}}
+          ${{env.activate_venv_venv}}
           python -m pip install --upgrade pip
           python -m pip install -e .[test]
 
@@ -51,21 +57,28 @@ jobs:
         if: ${{ matrix.backend == 'pynfft' || env.ref_backend == 'pynfft' }}
         shell: bash
         run: |
+          ${{env.activate_venv}}
           python -m pip install "pynfft2>=1.4.3"
 
       - name: Install pynufft
         if: ${{ matrix.backend == 'pynufft-cpu' || env.ref_backend == 'pynufft-cpu' }}
-        run: python -m pip install pynufft
+        run: |
+          ${{env.activate_venv}}
+          python -m pip install pynufft
 
       - name: Install finufft
         if: ${{ matrix.backend == 'finufft'  || env.ref_backend == 'finufft'}}
         shell: bash
-        run: python -m pip install finufft
+        run: |
+          ${{env.activate_venv}}
+          python -m pip install finufft
 
       - name: Install Sigpy
         if: ${{ matrix.backend == 'sigpy'  || env.ref_backend == 'sigpy'}}
         shell: bash
-        run: python -m pip install sigpy
+        run: |
+          ${{env.activate_venv}}
+          python -m pip install sigpy
 
       - name: Install BART
         if: ${{ matrix.backend == 'bart'  || env.ref_backend == 'bart'}}
@@ -81,12 +94,15 @@ jobs:
 
       - name: Install torchkbnufft-cpu
         if: ${{ matrix.backend == 'torchkbnufft-cpu'  || env.ref_backend == 'torchkbnufft-cpu'}}
-        run: python -m pip install torchkbnufft
+        run: |
+          ${{env.activate_venv}}
+          python -m pip install torchkbnufft
 
 
       - name: Run Tests
         shell: bash
         run: |
+          ${{env.activate_venv}}
           export COVERAGE_FILE=coverage_${{ matrix.backend }}
           pytest -k='operators' --backend ${{ matrix.backend }} --ref ${{ env.ref_backend }} --cov --disable-pytest-warnings --cov-branch --cov-report=term
       - name: Upload coverage
@@ -111,10 +127,8 @@ jobs:
       - name: Install mri-nufft and finufft
         shell: bash
         run: |
-          cd $RUNNER_WORKSPACE
-          python --version
-          python -m venv venv
-          source $RUNNER_WORKSPACE/venv/bin/activate
+          ${{ env.create_venv }}
+          ${{ env.activate_venv }}
           pip install --upgrade pip wheel
           pip install -e mri-nufft[test]
           pip install cupy-cuda12x finufft "numpy<2.0"
@@ -123,16 +137,14 @@ jobs:
         shell: bash
         if: ${{ matrix.backend != 'tensorflow'}}
         run: |
-          source $RUNNER_WORKSPACE/venv/bin/activate
+          source $RUNNER_WORKSPACE/venv/binactivate_venv
           pip install torch
 
[…]
