+++ .github/workflows/test-ci.yml
@@ -16,7 +16,7 @@ env:
   PYTHON_VERSION: "3.10"
   BART_VERSION: "0.8.00"
   ref_backend: "finufft"
-  create_venv: "python -m venv $RUNNER_WORKSPACE/venv/"
+  create_venv: "python -m venv $RUNNER_WORKSPACE/venv/ --clear"
   activate_venv_venv: "source $RUNNER_WORKSPACE/venv/bin/activate"
   setup_cuda: "export CUDA_BIN_PATH=/usr/local/cuda-12.1/;export PATH=/usr/local/cuda-12.1/bin/:${PATH};export LD_LIBRARY_PATH=/usr/local/cuda-12.1/lib64/:${LD_LIBRARY_PATH}"
 
@@ -104,7 +104,7 @@ jobs:
         run: |
           ${{env.activate_venv}}
           export COVERAGE_FILE=coverage_${{ matrix.backend }}
-          pytest -k='operators' --backend ${{ matrix.backend }} --ref ${{ env.ref_backend }} --cov --disable-pytest-warnings --cov-branch --cov-report=term
+          python -m pytest -k='operators' --backend ${{ matrix.backend }} --ref ${{ env.ref_backend }} --cov --disable-pytest-warnings --cov-branch --cov-report=term
       - name: Upload coverage
         uses: actions/upload-artifact@v4
         with:
@@ -129,16 +129,16 @@ jobs:
         run: |
           ${{ env.create_venv }}
           ${{ env.activate_venv }}
-          pip install --upgrade pip wheel
-          pip install -e mri-nufft[test]
-          pip install cupy-cuda12x finufft "numpy<2.0"
+          python -m pip install --upgrade pip wheel
+          python -m pip install -e mri-nufft[test]
+          python -m pip install cupy-cuda12x finufft "numpy<2.0"
 
       - name: Install torch with CUDA 12.1
         shell: bash
         if: ${{ matrix.backend != 'tensorflow'}}
         run: |
           source $RUNNER_WORKSPACE/venv/binactivate_venv
-          pip install torch
+          python -m pip install torch
 
       - name: Install backend
         shell: bash
@@ -146,13 +146,13 @@ jobs:
           ${{ env.activate_venv }}
           ${{ env.setup_cuda }}
           if [[ ${{ matrix.backend }} == "torchkbnufft-gpu" ]]; then
-            pip install torchkbnufft
+            python -m pip install torchkbnufft
           elif [[ ${{ matrix.backend }} == "tensorflow" ]]; then
-            pip install tensorflow-mri==0.21.0 tensorflow-probability==0.17.0 tensorflow-io==0.27.0 matplotlib==3.7
+            python -m pip install tensorflow-mri==0.21.0 tensorflow-probability==0.17.0 tensorflow-io==0.27.0 matplotlib==3.7
           elif [[ ${{ matrix.backend }} == "cufinufft" ]]; then
-            pip install "cufinufft<2.3" 
+            python -m pip install "cufinufft<2.3"
           else
-            pip install ${{ matrix.backend }}
+            python -m pip install ${{ matrix.backend }}
           fi
 
       - name: Run Tests
