### 2025-03-03 0fa51a2
Fix CI (#243)
* fix: use venv in all CI.

* more isolation.

* update ci
