+++ .github/workflows/test-ci.yml
@@ -104,7 +104,7 @@ jobs:
         run: |
           ${{ env.activate_venv }}
           export COVERAGE_FILE=coverage_${{ matrix.backend }}
-          python -m pytest -k='operators' --backend ${{ matrix.backend }} --ref ${{ env.ref_backend }} --cov --disable-pytest-warnings --cov-branch --cov-report=term
+          python -m pytest -k='operators' --backend ${{ matrix.backend }} --ref ${{ env.ref_backend }} --cov --disable-pytest-warnings --cov-branch --cov-report=term --durations=10
       - name: Upload coverage
         uses: actions/upload-artifact@v4
         with:
@@ -161,7 +161,7 @@ jobs:
           ${{ env.activate_venv }}
           cd $RUNNER_WORKSPACE/mri-nufft
           export COVERAGE_FILE=coverage_${{ matrix.backend }}
-          python -m pytest -k='operators' --ref ${{ env.ref_backend }} --backend ${{ matrix.backend }} --disable-pytest-warnings --cov --cov-branch --cov-report=term
+          python -m pytest -k='operators' --ref ${{ env.ref_backend }} --backend ${{ matrix.backend }} --disable-pytest-warnings --cov --cov-branch --cov-report=term --durations=10
 
       - name: Upload coverage
         if: success()
@@ -179,7 +179,7 @@ jobs:
           rm -rf finufft
           rm -rf gpuNUFFT
           rm -rf venv
- 
+
   get-commit-message:
     runs-on: ubuntu-latest
     outputs:
@@ -194,7 +194,7 @@ jobs:
             COMMIT_MESSAGE=$(git log -1 --pretty=%B)
             echo "Commit message $COMMIT_MESSAGE"
             echo "::set-output name=message::$COMMIT_MESSAGE"
-    
+
   test-examples:
     runs-on: gpu
     needs: get-commit-message
@@ -222,7 +222,7 @@ jobs:
           python -m pip install --upgrade pip
           python -m pip install -e .[extra,test,dev]
           python -m pip install finufft pooch brainweb-dl torch fastmri
-        
+
       - name: Install GPU related interfaces
         run: |
           ${{ env.activate_venv }}
@@ -235,7 +235,7 @@ jobs:
         run: |
           ${{ env.activate_venv }}
           export COVERAGE_FILE=coverage_plots
-          pytest examples tests -k="not operators" --cov --cov-branch --cov-report=term
+          pytest examples tests -k="not operators" --cov --cov-branch --cov-report=term --durations=10
 
       - name: Cache brainweb-dl directory
         uses: actions/cache/save@v4
@@ -270,7 +270,7 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: ${{ env.PYTHON_VERSION }}
-          
+
       - name: add the coverage tool
         shell: bash
         run: |
@@ -320,7 +320,7 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-      
+
       - name: Get history and tags for SCM versioning to work
         run: |
           git fetch --prune --unshallow
@@ -330,7 +330,7 @@ jobs:
         uses: actions/setup-python@v5
         with:
           python-version: "3.10"
-        
+
       - name: Install dependencies
         shell: bash -l {0}
         run: |
@@ -339,23 +339,23 @@ jobs:
           python -m pip install --upgrade pip
           python -m pip install .[doc]
           python -m pip install finufft
-      
+
       - name: Install GPU related interfaces
         run: |
           ${{ env.activate_venv }}
           ${{ env.setup_cuda }}
           pip install cupy-cuda12x torch
           python -m pip install gpuNUFFT "cufinufft<2.3"
-      
+
       - name: Build API documentation
         run: |
[…]
