### 2025-03-03 7d60349
Add gallery examples for ECCENTRIC trajectories
