### 2025-03-04 4fbe2e9
test: extract non operators test to a dedicated job.
