### 2025-03-04 8528b45
tweak test examples
