+++ .github/workflows/test-ci.yml
@@ -282,7 +282,7 @@ jobs:
         run: |
           ${{ env.activate_venv }}
           export COVERAGE_FILE=coverage_plots
-          pytest examples tests -n=4 -k="not operators" --cov --cov-branch --cov-report=term --durations=10
+          python -m pytest examples -n=4 --cov --cov-branch --cov-report=term --durations=10
 
       - name: Cache brainweb-dl directory
         uses: actions/cache/save@v4
