### 2025-03-04 32be4f6
feat: optimize CI runs (#245)
* report durations for tests.

* tests: use smaller array in 3D for off-resonance.

* feat: slim down tests cases for stacked nufft.

* test: extract non operators test to a dedicated job.

* feat: run examples in parallel.

* feat: disable expensive test.

* tweak test examples

* faster example

* fix
