+++ _sources/generated/autoexamples/GPU/example_learn_samples.rst
@@ -211,9 +211,9 @@ Setup data
 
  .. code-block:: none
 
-    /volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/GPU/example_learn_samples.py:117: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at ../aten/src/ATen/native/Copy.cpp:308.)
+    /volatile/github-ci-mind-inria/gpu_mind_runner/_work/mri-nufft/mri-nufft/examples/GPU/example_learn_samples.py:117: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at /pytorch/aten/src/ATen/native/Copy.cpp:308.)
       mri_2D = torch.Tensor(np.flipud(bwdl.get_mri(4, "T1")[80, ...]).astype(np.complex64))[
-    /volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/GPU/example_learn_samples.py:84: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
+    /volatile/github-ci-mind-inria/gpu_mind_runner/_work/mri-nufft/mri-nufft/examples/GPU/example_learn_samples.py:84: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
       axs[0].imshow(np.abs(mri_2D[0]), cmap="gray")
 
 
@@ -282,7 +282,7 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.29947168]      1%|          | 1/100 [00:00<01:15,  1.31steps/s, loss=0.29947168]      1%|          | 1/100 [00:00<01:15,  1.31steps/s, loss=0.38006115]      2%|▏         | 2/100 [00:01<01:07,  1.45steps/s, loss=0.38006115]      2%|▏         | 2/100 [00:01<01:07,  1.45steps/s, loss=0.30661485]      3%|▎         | 3/100 [00:02<01:07,  1.43steps/s, loss=0.30661485]      3%|▎         | 3/100 [00:02<01:07,  1.43steps/s, loss=0.2776927]       4%|▍         | 4/100 [00:02<01:06,  1.44steps/s, loss=0.2776927]      4%|▍         | 4/100 [00:02<01:06,  1.44steps/s, loss=0.27064103]      5%|▌         | 5/100 [00:03<01:09,  1.37steps/s, loss=0.27064103]      5%|▌         | 5/100 [00:03<01:09,  1.37steps/s, loss=0.25571805]      6%|▌         | 6/100 [00:04<01:06,  1.41steps/s, loss=0.25571805]      6%|▌         | 6/100 [00:04<01:06,  1.41steps/s, loss=0.24902384]      7%|▋         | 7/100 [00:04<01:03,  1.47steps/s, loss=0.24902384]      7%|▋         | 7/100 [00:04<01:03,  1.47steps/s, loss=0.24352454]      8%|▊         | 8/100 [00:05<01:02,  1.47steps/s, loss=0.24352454]      8%|▊         | 8/100 [00:05<01:02,  1.47steps/s, loss=0.236987]        9%|▉         | 9/100 [00:06<01:01,  1.49
[…]
