+++ _sources/generated/autoexamples/example_trajectory_tools.rst
@@ -2135,7 +2135,7 @@ References
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 52.975 seconds)
+   **Total running time of the script:** (0 minutes 51.651 seconds)
 
 
 .. _sphx_glr_download_generated_autoexamples_example_trajectory_tools.py:
