+++ _sources/generated/autoexamples/GPU/example_learn_straight_line_readouts.rst
@@ -37,12 +37,12 @@ The image resolution is kept small to reduce computation time.
 
    !pip install mri-nufft[gpunufft]
 
-.. GENERATED FROM PYTHON SOURCE LINES 24-26
+.. GENERATED FROM PYTHON SOURCE LINES 23-25
 
 Imports
 -------
 
-.. GENERATED FROM PYTHON SOURCE LINES 26-39
+.. GENERATED FROM PYTHON SOURCE LINES 25-38
 
 .. code-block:: Python
 
@@ -66,7 +66,7 @@ Imports
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 40-45
+.. GENERATED FROM PYTHON SOURCE LINES 39-44
 
 Setup a simple class to learn trajectory
 ----------------------------------------
@@ -74,7 +74,7 @@ Setup a simple class to learn trajectory
     While we are only learning the NUFFT operator, we still need the gradient `wrt_data=True` to have all the gradients computed correctly.
     See [Projector]_ for more details.
 
-.. GENERATED FROM PYTHON SOURCE LINES 45-115
+.. GENERATED FROM PYTHON SOURCE LINES 44-114
 
 .. code-block:: Python
 
@@ -155,12 +155,12 @@ Setup a simple class to learn trajectory
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 116-118
+.. GENERATED FROM PYTHON SOURCE LINES 115-117
 
 Util function to plot the state of the model
 --------------------------------------------
 
-.. GENERATED FROM PYTHON SOURCE LINES 118-163
+.. GENERATED FROM PYTHON SOURCE LINES 117-162
 
 .. code-block:: Python
 
@@ -176,18 +176,18 @@ Util function to plot the state of the model
         axs[0].axis("off")
         axs[0].set_title("MR Image")
         if traj.shape[-1] == 3:
-            if i is not None and i > 50:
+            if i is not None and i > 20:
                 axs[1].scatter(*traj.T[1:3, 0], s=10, color="blue")
             else:
                 fig_kwargs = {}
                 plt_kwargs = {"s": 1, "alpha": 0.2}
                 if i is not None:
                     fig_kwargs["azim"], fig_kwargs["elev"] = (
-                        i / 50 * 60 - 60,
-                        30 - i / 50 * 30,
+                        i / 25 * 60 - 60,
+                        30 - i / 25 * 30,
                     )
-                    plt_kwargs["alpha"] = 0.2 + 0.8 * i / 50
-                    plt_kwargs["s"] = 1 + 9 * i / 50
+                    plt_kwargs["alpha"] = 0.2 + 0.8 * i / 20
+                    plt_kwargs["s"] = 1 + 9 * i / 20
                 axs[1].remove()
                 axs[1] = fig.add_subplot(*fig_grid, 2, projection="3d", **fig_kwargs)
                 for shot in traj:
@@ -216,12 +216,12 @@ Util function to plot the state of the model
 
 
 
-.. GENERATED FROM PYTHON SOURCE LINES 164-166
+.. GENERATED FROM PYTHON SOURCE LINES 163-165
 
 Setup model and optimizer
 -------------------------
 
[…]
