### 2025-03-10 c616f88
Explicit the ECCENTRIC acronym
