### 2025-03-12 29605d2
Clean signal gradient display (#246)
