+++ _sources/generated/autoexamples/example_gif_2D.rst
@@ -274,7 +274,7 @@ Animation rendering
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (1 minutes 18.132 seconds)
+   **Total running time of the script:** (1 minutes 22.562 seconds)
 
 
 .. _sphx_glr_download_generated_autoexamples_example_gif_2D.py:
