### 2025-03-16 #7743 @Kevin-Mattheus-Moerman
@bwheelz36 thanks for your help also, you have one box unticked on community guidelines. The project does have these (https://github.com/mind-inria/mri-nufft/blob/master/CONTRIBUTING.md) let me know if you are happy to tick that last box and if you are done with your review at this point. Thanks!
