### 2025-03-17 79e25a3
Fix Lloyd's sampling coordinates
