### 2025-03-17 68fdec4
fixup! Fix Lloyd's sampling coordinates
