+++ _sources/generated/autoexamples/example_learn_samples_multires.rst
@@ -462,10 +462,10 @@ Training loop
       fig, axs = plt.subplots(2, 2, figsize=(10, 10), num=1)
     /volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/example_learn_samples_multires.py:136: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
       axs[0].imshow(np.abs(image[0]), cmap="gray")
-      3%|▎         | 1/30 [00:00<00:14,  2.04steps/s, loss=(array(0.86704654, dtype=float32),)]      3%|▎         | 1/30 [00:00<00:14,  2.04steps/s, loss=(array(0.5286128, dtype=float32),)]       7%|▋         | 2/30 [00:00<00:13,  2.09steps/s, loss=(array(0.5286128, dtype=float32),)]      7%|▋         | 2/30 [00:00<00:13,  2.09steps/s, loss=(array(0.36503774, dtype=float32),)]     10%|█         | 3/30 [00:01<00:12,  2.12steps/s, loss=(array(0.36503774, dtype=float32),)]     10%|█         | 3/30 [00:01<00:12,  2.12steps/s, loss=(array(0.4146319, dtype=float32),)]      13%|█▎        | 4/30 [00:01<00:12,  2.12steps/s, loss=(array(0.4146319, dtype=float32),)]     13%|█▎        | 4/30 [00:01<00:12,  2.12steps/s, loss=(array(0.30785707, dtype=float32),)]     17%|█▋        | 5/30 [00:02<00:12,  2.05steps/s, loss=(array(0.30785707, dtype=float32),)]     17%|█▋        | 5/30 [00:02<00:12,  2.05steps/s, loss=(array(0.32219106, dtype=float32),)]     20%|██        | 6/30 [00:02<00:11,  2.08steps/s, loss=(array(0.32219106, dtype=float32),)]     20%|██        | 6/30 [00:02<00:11,  2.08steps/s, loss=(array(0.34580207, dtype=float32),)]     23%|██▎       | 7/30 [00:03<00:12,  1.82steps/s, loss=(array(0.34580207, dtype=float32),)]     23%|██▎       | 7/30 [00:03<00:12,  1.82steps/s, loss=(array(0.3096822, dtype=float32),)]      27%|██▋       | 8/30 [00:04<00:11,  1.92steps/s, loss=(array(0.3096822, dtype=float32),)]     27%|██▋       | 8/30 [00:04<00:11,  1.92steps/s, loss=(array(0.29107636, dtype=float32),)]     30%|███       | 9/30 [00:04<00:10,  1.99steps/s, loss=(array(0.29107636, dtype=float32),)]     30%|███       | 9/30 [00:04<00:10,  1.99steps/s, loss=(array(0.28929195, dtype=float32),)]     33%|███▎      | 10/30 [00:04<00:09,  2.04steps/s, loss=(array(0.28929195, dtype=float32),)]     33%|███▎      | 10/30 [00:04<00:09,  2.04steps/s, loss=(array(0.27598783, dtype=float32),)]     37%|███▋      | 11/30 [00:05<00:09,  2.05steps/s, loss=(array(0.27598783, dtype=float32),)]     37%|███▋      | 11/30 [00:05<00:09,  2.05steps/s, loss=(array(0.27360773, dtype=float32),)]     40%|████      | 12/30 [00:05<00:08,  2.05steps/s, loss=(array(0.27360773, dtype=float32),)]     40%|████      | 12/30 [00:05<00:08,  2.05steps
[…]
