+++ _sources/generated/autoexamples/GPU/example_learn_samples_multicoil.rst
@@ -334,7 +334,7 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466064]      1%|          | 1/100 [00:00<01:20,  1.23steps/s, loss=0.15466064]      1%|          | 1/100 [00:01<01:20,  1.23steps/s, loss=0.27207476]      2%|▏         | 2/100 [00:02<01:51,  1.13s/steps, loss=0.27207476]      2%|▏         | 2/100 [00:02<01:51,  1.13s/steps, loss=0.20695516]      3%|▎         | 3/100 [00:03<01:50,  1.14s/steps, loss=0.20695516]      3%|▎         | 3/100 [00:03<01:50,  1.14s/steps, loss=0.10875541]      4%|▍         | 4/100 [00:04<01:47,  1.12s/steps, loss=0.10875541]      4%|▍         | 4/100 [00:04<01:47,  1.12s/steps, loss=0.10863988]      5%|▌         | 5/100 [00:05<01:46,  1.12s/steps, loss=0.10863988]      5%|▌         | 5/100 [00:05<01:46,  1.12s/steps, loss=0.09188324]      6%|▌         | 6/100 [00:06<01:41,  1.08s/steps, loss=0.09188324]      6%|▌         | 6/100 [00:06<01:41,  1.08s/steps, loss=0.078636155]      7%|▋         | 7/100 [00:07<01:40,  1.08s/steps, loss=0.078636155]      7%|▋         | 7/100 [00:08<01:40,  1.08s/steps, loss=0.07397699]       8%|▊         | 8/100 [00:08<01:46,  1.16s/steps, loss=0.07397699]      8%|▊         | 8/100 [00:09<01:46,  1.16s/steps, loss=0.07340166]      9%|▉         | 9/100 [00:10<01:53,  1.25s/steps, loss=0.07340166]      9%|▉         | 9/100 [00:10<01:53,  1.25s/steps, loss=0.06760576]     10%|█         | 10/100 [00:11<01:48,  1.20s/steps, loss=0.06760576]     10%|█         | 10/100 [00:11<01:48,  1.20s/steps, loss=0.06612788]     11%|█         | 11/100 [00:12<01:42,  1.15s/steps, loss=0.06612788]     11%|█         | 11/100 [00:12<01:42,  1.15s/steps, loss=0.061898775]     12%|█▏        | 12/100 [00:13<01:37,  1.11s/steps, loss=0.061898775]     12%|█▏        | 12/100 [00:14<01:37,  1.11s/steps, loss=0.058699183]     13%|█▎        | 13/100 [00:14<01:43,  1.19s/steps, loss=0.058699183]     13%|█▎        | 13/100 [00:15<01:43,  1.19s/steps, loss=0.059690844]     14%|█▍        | 14/100 [00:16<01:44,  1.22s/steps, loss=0.059690844]     14%|█▍        | 14/100 [00:16<01:44,  1.22s/steps, loss=0.05754113]      15%|█▌        | 15/100 [00:17<01:43,  1.22s/steps, loss=0.05754113]     15%|█▌        | 15/100 [00:17<01:43,  1.22s/steps, loss=0.051239453]     16%|█▌        | 16/100 [00:18<01:39,  1.19s/steps, loss=0.051239453]     16%|█▌        | 16/100 [00:19<01:39,  1.19s/steps, loss=0.04995346]      17%|█▋        | 17/100 [00:19<01:38,  1.18s/steps, loss=0.04995346]     17%|█▋        | 17/100 [
[…]
