### 2025-03-22 mind-inria/mri-nufft#250 @cncastillo
JOSS review
Hi! Congratulations on the excellent package. I’m sure it will be very useful for the community. These are some comments regarding: https://github.com/openjournals/joss-reviews/issues/7743

I have a list of minor issues:

Manuscript

- [ ] Missing citation for FINUFFT
- [ ] LaTeX error on line 81

README

- [ ] Make the first example bulletproof

I encountered some issues when running [this example](https://github.com/mind-inria/mri-nufft?tab=readme-ov-file#usage). It did not work after following the installation instructions because the FINUFFT backend was not installed.

I strongly believe that a user's first experience with your software should be flawless. While the documentation explains backend installation well, I recommend making the README example self-contained:

```python
# Install backend first: pip install mri-nufft[finufft]
NufftOperator = mrinufft.get_operator("finufft")
```

Cheers!
