### 2025-03-22 #7743 @cncastillo
I think this package is a great example of an open-source project following good coding practices. They include tests, code-coverage, docs (with separate sections for tutorials, explanations, etc), versioned releases, and more. Lovely to see software like this, and the quality and effort put into it really shines through. Congratulations to the package authors, as I know how long can this take!

- Minor issues: https://github.com/mind-inria/mri-nufft/issues/250

Cheers!
