### 2025-03-22 d997125
fix(joss): missing dollar sign
