+++ _sources/generated/autoexamples/GPU/example_fastMRI_UNet.rst
@@ -386,8 +386,6 @@ Get the image on which we will train our U-Net Model
 
  .. code-block:: none
 
-    /volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/GPU/example_fastMRI_UNet.py:131: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at /pytorch/aten/src/ATen/native/Copy.cpp:308.)
-      mri_2D = torch.Tensor(np.flipud(bwdl.get_mri(4, "T1")[80, ...]).astype(np.complex64))[
     /volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/GPU/example_fastMRI_UNet.py:104: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
       axs[0].imshow(np.abs(mri_2D[0]), cmap="gray")
 
@@ -456,9 +454,7 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=1]/volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/GPU/example_fastMRI_UNet.py:104: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
-      axs[0].imshow(np.abs(mri_2D[0]), cmap="gray")
-      1%|          | 1/100 [00:00<01:30,  1.09steps/s, loss=1]      1%|          | 1/100 [00:01<01:30,  1.09steps/s, loss=0.903]      2%|▏         | 2/100 [00:01<01:25,  1.15steps/s, loss=0.903]      2%|▏         | 2/100 [00:01<01:25,  1.15steps/s, loss=0.829]      3%|▎         | 3/100 [00:02<01:33,  1.04steps/s, loss=0.829]      3%|▎         | 3/100 [00:02<01:33,  1.04steps/s, loss=0.778]      4%|▍         | 4/100 [00:03<01:27,  1.10steps/s, loss=0.778]      4%|▍         | 4/100 [00:03<01:27,  1.10steps/s, loss=0.741]      5%|▌         | 5/100 [00:04<01:25,  1.12steps/s, loss=0.741]      5%|▌         | 5/100 [00:04<01:25,  1.12steps/s, loss=0.714]      6%|▌         | 6/100 [00:05<01:24,  1.12steps/s, loss=0.714]      6%|▌         | 6/100 [00:05<01:24,  1.12steps/s, loss=0.678]      7%|▋         | 7/100 [00:06<01:24,  1.10steps/s, loss=0.678]      7%|▋         | 7/100 [00:06<01:24,  1.10steps/s, loss=0.645]      8%|▊         | 8/100 [00:07<01:24,  1.09steps/s, loss=0.645]      8%|▊         | 8/100 [00:07<01:24,  1.09steps/s, loss=0.616]      9%|▉         | 9/100 [00:08<01:23,  1.08steps/s, loss=0.616]      9%|▉         | 9/100 [00:08<01:23,  1.08steps/s, loss=0.592]     10%|█         | 10/100 [00:09<01:22,  1.09steps/s, loss=0.592]     10%|█         | 10/100 [00:09<01:22, 
[…]
