+++ _sources/generated/autoexamples/GPU/example_learn_samples_multicoil.rst
@@ -334,7 +334,9 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=0.15466064]      1%|          | 1/100 [00:00<01:19,  1.25steps/s, loss=0.15466064]      1%|          | 1/100 [00:01<01:19,  1.25steps/s, loss=0.2716676]       2%|▏         | 2/100 [00:01<01:20,  1.22steps/s, loss=0.2716676]      2%|▏         | 2/100 [00:01<01:20,  1.22steps/s, loss=0.20653139]      3%|▎         | 3/100 [00:02<01:19,  1.22steps/s, loss=0.20653139]      3%|▎         | 3/100 [00:02<01:19,  1.22steps/s, loss=0.10851698]      4%|▍         | 4/100 [00:03<01:21,  1.18steps/s, loss=0.10851698]      4%|▍         | 4/100 [00:03<01:21,  1.18steps/s, loss=0.10890454]      5%|▌         | 5/100 [00:04<01:19,  1.19steps/s, loss=0.10890454]      5%|▌         | 5/100 [00:04<01:19,  1.19steps/s, loss=0.09083839]      6%|▌         | 6/100 [00:05<01:19,  1.18steps/s, loss=0.09083839]      6%|▌         | 6/100 [00:05<01:19,  1.18steps/s, loss=0.07860584]      7%|▋         | 7/100 [00:05<01:17,  1.20steps/s, loss=0.07860584]      7%|▋         | 7/100 [00:06<01:17,  1.20steps/s, loss=0.073942274]      8%|▊         | 8/100 [00:06<01:16,  1.21steps/s, loss=0.073942274]      8%|▊         | 8/100 [00:06<01:16,  1.21steps/s, loss=0.07300216]       9%|▉         | 9/100 [00:07<01:21,  1.12steps/s, loss=0.07300216]      9%|▉         | 9/100 [00:07<01:21,  1.12steps/s, loss=0.06790188]     10%|█         | 10/100 [00:08<01:17,  1.16steps/s, loss=0.06790188]     10%|█         | 10/100 [00:08<01:17,  1.16steps/s, loss=0.06560062]     11%|█         | 11/100 [00:09<01:14,  1.20steps/s, loss=0.06560062]     11%|█         | 11/100 [00:09<01:14,  1.20steps/s, loss=0.0618926]      12%|█▏        | 12/100 [00:10<01:11,  1.23steps/s, loss=0.0618926]     12%|█▏        | 12/100 [00:10<01:11,  1.23steps/s, loss=0.057889104]     13%|█▎        | 13/100 [00:10<01:10,  1.24steps/s, loss=0.057889104]     13%|█▎        | 13/100 [00:11<01:10,  1.24steps/s, loss=0.05963811]      14%|█▍        | 14/100 [00:11<01:11,  1.21steps/s, loss=0.05963811]     14%|█▍        | 14/100 [00:11<01:11,  1.21steps/s, loss=0.057414204]     15%|█▌        | 15/100 [00:12<01:09,  1.23steps/s, loss=0.057414204]     15%|█▌        | 15/100 [00:12<01:09,  1.23steps/s, loss=0.052272487]     16%|█▌        | 16/100 [00:13<01:07,  1.24steps/s, loss=0.052272487]     16%|█▌        | 16/100 [00:13<01:07,  1.24steps/s, loss=0.049586616]     17%|█▋        | 17/100 [00:14<01:06,  1.25steps/s, loss=0.049586616]     17%|█▋        | 17/100 [00:14<01:0
[…]
