### 2025-03-23 366192c
fix(joss): Address #250 (#251)
* fix(joss): add finufft-refs

* fix(joss): update ref in paper

* update example in readme
