### 2025-03-23 mind-inria/mri-nufft#250 @paquiteau
Hi @cncastillo

Thank you for your feedback ! I updated the README and corrected the paper :)
