+++ _sources/generated/autoexamples/GPU/example_fastMRI_UNet.rst
@@ -386,7 +386,9 @@ Get the image on which we will train our U-Net Model
 
  .. code-block:: none
 
-    /volatile/github-ci-mind-inria/gpu_runner2/_work/mri-nufft/mri-nufft/examples/GPU/example_fastMRI_UNet.py:104: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
+    /volatile/github-ci-mind-inria/gpu_mind_runner/_work/mri-nufft/mri-nufft/examples/GPU/example_fastMRI_UNet.py:131: UserWarning: Casting complex values to real discards the imaginary part (Triggered internally at /pytorch/aten/src/ATen/native/Copy.cpp:308.)
+      mri_2D = torch.Tensor(np.flipud(bwdl.get_mri(4, "T1")[80, ...]).astype(np.complex64))[
+    /volatile/github-ci-mind-inria/gpu_mind_runner/_work/mri-nufft/mri-nufft/examples/GPU/example_fastMRI_UNet.py:104: DeprecationWarning: __array_wrap__ must accept context and return_scalar arguments (positionally) in the future. (Deprecated NumPy 2.0)
       axs[0].imshow(np.abs(mri_2D[0]), cmap="gray")
 
 
@@ -454,7 +456,9 @@ Start training loop
 
  .. code-block:: none
 
-      0%|          | 0/100 [00:00<?, ?steps/s]      0%|          | 0/100 [00:00<?, ?steps/s, loss=1.06]      1%|          | 1/100 [00:01<02:38,  1.60s/steps, loss=1.06]      1%|          | 1/100 [00:01<02:38,  1.60s/steps, loss=0.97]      2%|▏         | 2/100 [00:02<02:18,  1.41s/steps, loss=0.97]      2%|▏         | 2/100 [00:03<02:18,  1.41s/steps, loss=0.899]      3%|▎         | 3/100 [00:04<02:18,  1.43s/steps, loss=0.899]      3%|▎         | 3/100 [00:04<02:18,  1.43s/steps, loss=0.84]       4%|▍         | 4/100 [00:06<02:32,  1.59s/steps, loss=0.84]      4%|▍         | 4/100 [00:06<02:32,  1.59s/steps, loss=0.793]      5%|▌         | 5/100 [00:08<02:53,  1.82s/steps, loss=0.793]      5%|▌         | 5/100 [00:08<02:53,  1.82s/steps, loss=0.757]      6%|▌         | 6/100 [00:10<02:56,  1.87s/steps, loss=0.757]      6%|▌         | 6/100 [00:10<02:56,  1.87s/steps, loss=0.711]      7%|▋         | 7/100 [00:12<03:01,  1.95s/steps, loss=0.711]      7%|▋         | 7/100 [00:12<03:01,  1.95s/steps, loss=0.671]      8%|▊         | 8/100 [00:14<02:52,  1.87s/steps, loss=0.671]      8%|▊         | 8/100 [00:14<02:52,  1.87s/steps, loss=0.637]      9%|▉         | 9/100 [00:15<02:34,  1.70s/steps, loss=0.637]      9%|▉         | 9/100 [00:15<02:34,  1.70s/steps, loss=0.608]     10%|█         | 10/100 [00:17<02:27,  1.64s/steps, loss=0.608]     10%|█         | 10/100 [00:17<02:27,  1.64s/steps, loss=0.582]     11%|█         
[…]
