### 2025-03-25 1c4c18e
class method cg
