+++ examples/GPU/example_cg.py
@@ -10,18 +10,9 @@ for solving systems of linear equations of the form Ax = b, where A is a symmetr
 positive-definite matrix. The CG method is an iterative algorithm that is particularly
 useful for large, sparse systems where direct methods are computationally expensive.
 
-The Conjugate Gradient method is widely used in various scientific and engineering
-applications, including solving partial differential equations, optimization problems,
-and machine learning tasks.
+This method is inspired by techniques from [SigPy]_ and
+[Aquaulb's MOOC]_, as well as general knowledge in [Wikipedia Conjugate Gradient]_.
 
-References
-----------
-- Inpirations:
-        - https://sigpy.readthedocs.io/en/latest/_modules/sigpy/alg.html#ConjugateGradient
-        - https://aquaulb.github.io/book_solving_pde_mooc/solving_pde_mooc/notebooks/05_IterativeMethods/05_02_Conjugate_Gradient.html
-- Wikipedia:
-        - https://en.wikipedia.org/wiki/Conjugate_gradient_method
-        - https://en.wikipedia.org/wiki/Momentum
 """
 
 # %%
@@ -100,3 +91,17 @@ plt.plot(kspace_data, label="acquired kspace")
 plt.plot(nufft.op(image), alpha=0.7,label="reconstructed kspace")
 plt.legend(loc="lower left", fontsize=8)
 # %%
+# References
+# ==========
+#
+# .. [SigPy] SigPy Documentation. Conjugate Gradient Method.
+#    https://sigpy.readthedocs.io/en/latest/_modules/sigpy/alg.html#ConjugateGradient
+
+# .. [Aquaulb] Aquaulb's MOOC: Solving PDE with Iterative Methods.
+#    https://aquaulb.github.io/book_solving_pde_mooc/solving_pde_mooc/notebooks/05_IterativeMethods/05_02_Conjugate_Gradient.html
+
+# .. [Wikipedia Conjugate Gradient] Wikipedia: Conjugate Gradient Method.
+#    https://en.wikipedia.org/wiki/Conjugate_gradient_method
+
+# .. [Wikipedia Momentum] Wikipedia: Momentum.
+#    https://en.wikipedia.org/wiki/Momentum
