+++ src/mrinufft/operators/base.py
@@ -373,7 +373,7 @@ class FourierOperatorBase(ABC):
         kspace_data: np.ndarray
             The k-space data to reconstruct.
         computer_loss: bool
-            Whether to compute the loss at each iteration. 
+            Whether to compute the loss at each iteration.
             If True, loss is calculated and returned, otherwise, it's skipped.
         **kwargs:
             Extra arguments to pass to the conjugate gradient method.
@@ -381,12 +381,14 @@ class FourierOperatorBase(ABC):
         Returns
         -------
         np.ndarray, np.ndarray
-            Reconstructed image and loss value if compute_loss is True. 
+            Reconstructed image and loss value if compute_loss is True.
         """
         from ..extras.gradient import cg
 
-        return cg(operator=self, kspace_data=kspace_data, compute_loss=compute_loss, **kwargs)
-    
+        return cg(
+            operator=self, kspace_data=kspace_data, compute_loss=compute_loss, **kwargs
+        )
+
     @property
     def uses_sense(self):
         """Return True if the operator uses sensitivity maps."""
