### 2025-03-25 6899afb
docstring of cg
