+++ src/mrinufft/operators/base.py
@@ -380,8 +380,10 @@ class FourierOperatorBase(ABC):
 
         Returns
         -------
-        np.ndarray, np.ndarray
-            Reconstructed image and loss value if compute_loss is True.
+        np.ndarray
+            Reconstructed image 
+        np.ndarray, optional
+            array of loss at each iteration, if compute_loss is True. 
         """
         from ..extras.gradient import cg
 
