+++ src/mrinufft/extras/gradient.py
@@ -64,6 +64,6 @@ def cg(operator, kspace_data, x_init=None, num_iter=10, tol=1e-4, compute_loss=F
         velocity = grad_new + beta * velocity
 
         image = image - velocity / Lipschitz_cst
-        if compute_loss:  
-             loss.append(calculate_loss(image))  
+        if compute_loss:
+            loss.append(calculate_loss(image))
     return image if loss is None else (image, np.array(loss))
