+++ src/mrinufft/extras/gradient.py
@@ -35,7 +35,7 @@ def cg(operator, kspace_data, x_init=None, num_iter=10, tol=1e-4, compute_loss=F
     image : numpy.ndarray
               The reconstructed image after the optimization process.
     """
-    Lipschitz_cst = operator.get_lipschitz_cst()
+    lipschitz_cst = operator.get_lipschitz_cst()
     image = (
         np.zeros(operator.shape, dtype=type(kspace_data[0]))
         if x_init is None
@@ -44,7 +44,7 @@ def cg(operator, kspace_data, x_init=None, num_iter=10, tol=1e-4, compute_loss=F
     velocity = np.zeros_like(image)
 
     grad = operator.data_consistency(image, kspace_data)
-    velocity = tol * velocity + grad / Lipschitz_cst
+    velocity = tol * velocity + grad / lipschitz_cst
     image = image - velocity
 
     def calculate_loss(image):
@@ -63,7 +63,7 @@ def cg(operator, kspace_data, x_init=None, num_iter=10, tol=1e-4, compute_loss=F
         beta = max(0, beta)  # Polak-Ribiere formula is used to compute the beta
         velocity = grad_new + beta * velocity
 
-        image = image - velocity / Lipschitz_cst
+        image = image - velocity / lipschitz_cst
         if compute_loss:
             loss.append(calculate_loss(image))
     return image if loss is None else (image, np.array(loss))
