+++ examples/GPU/example_cg.py
@@ -15,7 +15,7 @@ applications, including solving partial differential equations, optimization pro
 and machine learning tasks.
 
 This method is inspired by techniques from [SigPy]_ and
-[Aquaulb's MOOC]_, as well as general knowledge in [Wikipedia Conjugate Gradient]_.
+[Aquaulb]_'s MOOC, as well as general knowledge in [Wikipedia Conjugate Gradient]_.
 
 """
 
@@ -54,45 +54,28 @@ nufft = NufftOperator(
 
 # %%
 # Reconstruct the image using the CG method
-kspace_data = nufft.op(image)  # get the k-space data
+kspace_data = nufft.op(image) # get the k-space data
 reconstructed_image, loss = nufft.cg(
-    kspace_data=kspace_data, num_iter=50, compute_loss=True
-)  # reconstruct the image
+    kspace_data=kspace_data, x_init=nufft.adj_op(kspace_data).copy(), num_iter=50, compute_loss=True
+)  
 
 
 # Display the results
-def normalize(img, vmin, vmax):
-    return (img - vmin) / (vmax - vmin)
-
 
 plt.figure(figsize=(15, 10))
 plt.subplot(2, 3, 1)
 plt.title("Original image")
-plt.imshow(normalize(abs(image), abs(image).min(), abs(image).max()), cmap="gray")
+plt.imshow(image, cmap="gray")
 plt.colorbar()
 
 plt.subplot(2, 3, 2)
 plt.title("Conjugate gradient")
-plt.imshow(
-    normalize(
-        abs(reconstructed_image),
-        abs(reconstructed_image).min(),
-        abs(reconstructed_image).max(),
-    ),
-    cmap="gray",
-)
+plt.imshow(abs(reconstructed_image),vmin=image.min(),vmax=image.max(),cmap="gray")
 plt.colorbar()
 
 plt.subplot(2, 3, 3)
 plt.title("Adjoint NUFFT")
-plt.imshow(
-    normalize(
-        abs(nufft.adj_op(kspace_data)),
-        abs(nufft.adj_op(kspace_data)).min(),
-        abs(nufft.adj_op(kspace_data)).max(),
-    ),
-    cmap="gray",
-)
+plt.imshow(abs(nufft.adj_op(kspace_data)),vmin=image.min(),vmax=image.max(),cmap="gray")
 plt.colorbar()
 
 plt.subplot(2, 3, 4)
@@ -103,14 +86,14 @@ plt.grid()
 plt.subplot(2, 3, 5)
 plt.title("kspace from conjugate gradient")
 
-plt.plot(kspace_data, label="acquired kspace")
-plt.plot(nufft.op(reconstructed_image), alpha=0.7, label="reconstructed kspace")
+plt.plot(np.log(abs(kspace_data)), label="acquired kspace")
+plt.plot(np.log(abs(nufft.op(reconstructed_image))), alpha=0.7, label="reconstructed kspace")
 plt.legend(loc="lower left", fontsize=8)
 
 plt.subplot(2, 3, 6)
 plt.title("kspace from adjoint NUFFT")
-plt.plot(kspace_data, label="acquired kspace")
-plt.plot(nufft.op(image), alpha=0.7, label="reconstructed kspace")
+plt.plot(np.log(abs(kspace_data)), label="acquired kspace")
+plt.plot(np.log(abs(nufft.op(image))), alpha=0.7, label="reconstructed kspace")
 plt.legend(loc="lower left", fontsize=8)
 # %%
 # References
