### 2025-04-03 #7743 @Kevin-Mattheus-Moerman
@paquiteau please work on your points listed above. In addition can you check the below:
- [ ] Please see if you can address the missing DOI pointed out 👆
- [ ] Both `gpuUNUFFT` and `gpunufft`, as well as `CUFINUFFT` and `cufinufft` occur in the text. Please review these so projects are referred to consistently.
- [ ] The paper currently lacks an Acknowledgements section. This section is not a requirement, but please check if you need one e.g. to acknowledge funding.
Check the following typos:
- [ ] compatibily -> compatibility
- [ ] Traditionnaly -> Traditionally
- [ ] processus -> processes
- [ ] optionnaly -> optionally
- [ ] teslas -> Tesla
