### 2025-04-03 cbcdf2a
fix(joss): typos
