### 2025-04-03 #7743 @paquiteau
@Kevin-Mattheus-Moerman Thanks for you quick response !
I added the missing DOI and corrected the typos. I don't think we need an Acknowledgment section.
