### 2025-04-10 c9d3ad0
Fix ECCENTRIC docstrings
