### 2025-04-14 #7743 @Kevin-Mattheus-Moerman
@paquiteau

- [ ] Please edit the archive title to match the paper title.
- [ ] I have had to assign `v1.2.1-zenodo` as the version tag here instead of `v1.2.1`. Although the former has a release on your GitHub repository too, we need the archive tag and the release tag to match. Please let me know if you will edit the version tag for ZENODO to be `v1.2.1` (then I'll edit it here too) or if you are happy with the version tag `v1.2.1-zenodo`.
