### 2025-04-14 #7743 @paquiteau
Hello @Kevin-Mattheus-Moerman
I updated the title on zenodo. v1.2.1-zenodo and v1.2.1 are just two tags on the same commit, I had to do the zenodo one to trigger the archiving process.

I think we are (almost) done now, so I would like to thank you again for your investment and your patience in handling the review process.
