### 2025-04-15 #7743 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
